Median absolute deviation test for outliers

Performs a median absolute deviation (MAD) test for outliers.

mad_outliers(x, mask = !is.na(x), threshold = c(1.5, 3), return.score = FALSE)

Arguments

x
A vector of data.

mask
A logical mask that identifies a subgroup of x to compute non-parametric statistics on. Useful when a subset of quality-assured data is available. Default action is to ignore NA values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

Details

the values of threshold identify the multiplier of the median distance used to identify mild and extreme outliers. Default values are 1.5 for "mild" outliers and 3.0 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi     
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

mad_outliers(y)

#>   [1] mild outlier    mild outlier    extreme outlier not outlier    
#>   [5] not outlier     not outlier     extreme outlier mild outlier   
#>   [9] not outlier     not outlier     not outlier     not outlier    
#>  [13] not outlier     not outlier     mild outlier    not outlier    
#>  [17] not outlier     extreme outlier not outlier     not outlier    
#>  [21] not outlier     not outlier     mild outlier    not outlier    
#>  [25] not outlier     not outlier     extreme outlier extreme outlier
#>  [29] not outlier     extreme outlier not outlier     not outlier    
#>  [33] not outlier     not outlier     mild outlier    not outlier    
#>  [37] not outlier     extreme outlier mild outlier    not outlier    
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     mild outlier    mild outlier    mild outlier   
#>  [49] not outlier     not outlier     not outlier     not outlier    
#>  [53] not outlier     not outlier     not outlier     mild outlier   
#>  [57] not outlier     extreme outlier extreme outlier not outlier    
#>  [61] extreme outlier mild outlier    not outlier     mild outlier   
#>  [65] not outlier     not outlier     not outlier     not outlier    
#>  [69] not outlier     mild outlier    extreme outlier not outlier    
#>  [73] not outlier     not outlier     not outlier     not outlier    
#>  [77] extreme outlier mild outlier    not outlier     not outlier    
#>  [81] not outlier     extreme outlier not outlier     not outlier    
#>  [85] not outlier     mild outlier    not outlier     not outlier    
#>  [89] not outlier     not outlier     not outlier     not outlier    
#>  [93] mild outlier    not outlier     extreme outlier mild outlier   
#>  [97] not outlier     not outlier     not outlier     not outlier    
#> [101] not outlier     not outlier     not outlier     extreme outlier
#> [105] mild outlier    extreme outlier not outlier     not outlier    
#> [109] extreme outlier not outlier     not outlier     mild outlier   
#> [113] not outlier     not outlier     not outlier     not outlier    
#> [117] mild outlier    mild outlier    not outlier     not outlier    
#> [121] not outlier     extreme outlier not outlier     not outlier    
#> [125] not outlier     not outlier     mild outlier    not outlier    
#> [129] not outlier     not outlier     not outlier     not outlier    
#> [133] not outlier     not outlier     extreme outlier mild outlier   
#> [137] not outlier    
#> Levels: not outlier < mild outlier < extreme outlier

mad_outliers(y, mask)

#>   [1] extreme outlier extreme outlier extreme outlier not outlier    
#>   [5] extreme outlier extreme outlier extreme outlier mild outlier   
#>   [9] extreme outlier mild outlier    mild outlier    not outlier    
#>  [13] not outlier     not outlier     mild outlier    extreme outlier
#>  [17] not outlier     extreme outlier not outlier     extreme outlier
#>  [21] extreme outlier extreme outlier mild outlier    mild outlier   
#>  [25] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [29] not outlier     extreme outlier not outlier     not outlier    
#>  [33] mild outlier    not outlier     extreme outlier extreme outlier
#>  [37] not outlier     extreme outlier mild outlier    not outlier    
#>  [41] extreme outlier extreme outlier extreme outlier not outlier    
#>  [45] extreme outlier mild outlier    mild outlier    not outlier    
#>  [49] not outlier     extreme outlier mild outlier    extreme outlier
#>  [53] extreme outlier extreme outlier extreme outlier not outlier    
#>  [57] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [61] extreme outlier not outlier     not outlier     not outlier    
#>  [65] not outlier     extreme outlier extreme outlier not outlier    
#>  [69] not outlier     not outlier     extreme outlier extreme outlier
#>  [73] extreme outlier not outlier     extreme outlier mild outlier   
#>  [77] extreme outlier not outlier     extreme outlier not outlier    
#>  [81] not outlier     extreme outlier not outlier     not outlier    
#>  [85] not outlier     not outlier     extreme outlier not outlier    
#>  [89] not outlier     not outlier     not outlier     not outlier    
#>  [93] extreme outlier extreme outlier extreme outlier not outlier    
#>  [97] mild outlier    not outlier     mild outlier    extreme outlier
#> [101] not outlier     extreme outlier extreme outlier extreme outlier
#> [105] extreme outlier extreme outlier extreme outlier not outlier    
#> [109] extreme outlier not outlier     not outlier     not outlier    
#> [113] not outlier     extreme outlier mild outlier    extreme outlier
#> [117] extreme outlier mild outlier    extreme outlier extreme outlier
#> [121] extreme outlier extreme outlier mild outlier    not outlier    
#> [125] not outlier     extreme outlier mild outlier    extreme outlier
#> [129] not outlier     not outlier     not outlier     mild outlier   
#> [133] not outlier     not outlier     extreme outlier not outlier    
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

mad_outliers(y, mask, threshold = c(1, 2))

#>   [1] extreme outlier extreme outlier extreme outlier mild outlier   
#>   [5] extreme outlier extreme outlier extreme outlier mild outlier   
#>   [9] extreme outlier extreme outlier mild outlier    not outlier    
#>  [13] not outlier     not outlier     mild outlier    extreme outlier
#>  [17] not outlier     extreme outlier mild outlier    extreme outlier
#>  [21] extreme outlier extreme outlier mild outlier    mild outlier   
#>  [25] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [29] not outlier     extreme outlier not outlier     not outlier    
#>  [33] mild outlier    not outlier     extreme outlier extreme outlier
#>  [37] not outlier     extreme outlier mild outlier    not outlier    
#>  [41] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [45] extreme outlier mild outlier    mild outlier    mild outlier   
#>  [49] not outlier     extreme outlier mild outlier    extreme outlier
#>  [53] extreme outlier extreme outlier extreme outlier not outlier    
#>  [57] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [61] extreme outlier not outlier     not outlier     mild outlier   
#>  [65] not outlier     extreme outlier extreme outlier not outlier    
#>  [69] not outlier     mild outlier    extreme outlier extreme outlier
#>  [73] extreme outlier not outlier     extreme outlier mild outlier   
#>  [77] extreme outlier not outlier     extreme outlier not outlier    
#>  [81] not outlier     extreme outlier not outlier     not outlier    
#>  [85] not outlier     mild outlier    extreme outlier not outlier    
#>  [89] not outlier     not outlier     mild outlier    mild outlier   
#>  [93] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [97] extreme outlier mild outlier    mild outlier    extreme outlier
#> [101] not outlier     extreme outlier extreme outlier extreme outlier
#> [105] extreme outlier extreme outlier extreme outlier mild outlier   
#> [109] extreme outlier not outlier     mild outlier    not outlier    
#> [113] not outlier     extreme outlier mild outlier    extreme outlier
#> [117] extreme outlier mild outlier    extreme outlier extreme outlier
#> [121] extreme outlier extreme outlier mild outlier    not outlier    
#> [125] not outlier     extreme outlier mild outlier    extreme outlier
#> [129] not outlier     not outlier     not outlier     mild outlier   
#> [133] not outlier     mild outlier    extreme outlier mild outlier   
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

mad_outliers(y, return.score = TRUE)

#>   [1]   1.545901478   2.514831053  15.222941777  -0.519527268   0.767674399
#>   [6]   1.490175724  51.597684089  -1.997655843   0.102369199   0.014418325
#>  [11]           NaN  -0.798251806  -0.923834819  -1.000000000  -2.158090245
#>  [16]   0.363353462  -0.594140626  10.413778150  -0.322007129   1.212264645
#>  [21]   0.309289447   0.161432792  -2.184049954  -0.004313519   1.265342492
#>  [26]   1.388466634   3.383095848   5.358369537  -1.361890510 503.855927324
#>  [31]  -0.838775752  -1.370319719  -0.174985885  -0.677256277   2.594053022
#>  [36]   0.157786370  -1.365065571   3.644075552  -1.993521504  -0.729588282
#>  [41]   0.745780783   0.847951736   0.865384884  -0.519657001   0.097875412
#>  [46]  -2.001822167  -2.099065325  -1.786976675  -1.109573146   1.081995038
#>  [51]  -0.096776637   0.329512309   0.601803798   1.381319821   0.929884666
#>  [56]  -1.649570896   0.229417360   5.099217014  40.331519488   0.071565437
#>  [61]  10.491602088  -1.659549513  -1.069393396  -1.712799707  -0.823601298
#>  [66]   1.015803468   1.093575308  -0.551431273  -1.202104737  -1.873294430
#>  [71]  37.771685785   0.343224281   0.270860360  -0.630310462   0.194714195
#>  [76]  -0.003326621  24.019614101  -1.676075239   0.844506923  -0.568427901
#>  [81]  -1.403343737  12.755302322  -0.587543873  -0.729252347  -1.147027823
#>  [86]  -1.727067885   0.704826788  -1.130692534  -1.235271691  -0.792296862
#>  [91]  -0.394583674  -0.495269044   1.915165697   0.278667624  39.438637091
#>  [96]  -1.848266888   0.042174596  -0.283198782  -0.124631358   0.067021734
#> [101]  -1.415369394   0.133203699   0.326625767  12.767217300   1.632824775
#> [106]   5.750506929   0.070373968  -0.350533684  16.618696889  -1.161915134
#> [111]  -0.328253554  -1.587108001  -1.174948932   0.061641681  -0.169769867
#> [116]   0.243223427   1.545881463  -1.995147625   0.181009832   1.000000000
#> [121]   0.503452336  65.838089923  -0.095356046  -0.750351857  -1.285888605
#> [126]   0.576532945  -2.037701285   0.198000104  -1.359119665  -0.829516718
#> [131]  -0.562518464  -0.215154479  -1.283453312  -0.438316075   9.478129251
#> [136]  -1.831893022   1.183509955