Isolation Forest multivariate outlier detection

Performs outlier detection using an Isolation Forest.

moutlier_iforest(
  xs,
  mask = !Reduce("|", lapply(xs, is.na)),
  threshold = c(0.8, 0.9),
  return.score = FALSE,
  ...
)

Arguments

xs
A dataframe or list of vectors (which will be coerced to a numeric matrix).

mask
A logical vector that defines which values in x will used when computing statistics. Useful when a subset of quality-assured data is available. Default mask is non-NA Values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

...
Additional arguments to solitude::isolationForest$new(). note that the argument sample_size will be overwritten to use the number of unmasked data points, i.e. length(which(mask)).

Details

the values of threshold identify mild and extreme\ outliers based on the Isolation Forest score in the range [0,1]. Default values are 0.8 for "mild" outliers and 0.9 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y = sin(x) + noise
mask = noise < 1

if (requireNamespace("solitude", quietly = TRUE)) {
  moutlier_iforest(list(y))
  moutlier_iforest(list(x, y))
  moutlier_iforest(list(x, y), mask)
  moutlier_iforest(list(x, y), mask, threshold = c(1, 2))
  moutlier_iforest(list(x, y), return.score = TRUE)
}

#> INFO  [14:49:11.717] Building Isolation Forest ...  
#> INFO  [14:49:12.990] done 
#> INFO  [14:49:13.001] Computing depth of terminal nodes ...  
#> INFO  [14:49:13.845] done 
#> INFO  [14:49:13.859] Completed growing isolation forest 
#> INFO  [14:49:13.884] Building Isolation Forest ...  
#> INFO  [14:49:13.896] done 
#> INFO  [14:49:13.898] Computing depth of terminal nodes ...  
#> INFO  [14:49:14.200] done 
#> INFO  [14:49:14.212] Completed growing isolation forest 
#> INFO  [14:49:14.234] Building Isolation Forest ...  
#> INFO  [14:49:14.242] done 
#> INFO  [14:49:14.245] Computing depth of terminal nodes ...  
#> INFO  [14:49:14.563] done 
#> INFO  [14:49:14.573] Completed growing isolation forest 
#> INFO  [14:49:14.592] Building Isolation Forest ...  
#> INFO  [14:49:14.600] done 
#> INFO  [14:49:14.603] Computing depth of terminal nodes ...  
#> INFO  [14:49:14.856] done 
#> INFO  [14:49:14.864] Completed growing isolation forest 
#> INFO  [14:49:14.883] Building Isolation Forest ...  
#> INFO  [14:49:14.894] done 
#> INFO  [14:49:14.896] Computing depth of terminal nodes ...  
#> INFO  [14:49:15.163] done 
#> INFO  [14:49:15.176] Completed growing isolation forest 

#>   [1] 0.5967039 0.5745914 0.5601630 0.5794830 0.5580087 0.7507511 0.5444161
#>   [8] 0.5490508 0.5494741 0.5636272 0.5490508 0.5532984 0.5469393 0.5452559
#>  [15] 0.5524463 0.5465179 0.5448358 0.5524463 0.6459989 0.5482052 0.5541519
#>  [22] 0.5515955 0.5848668 0.5452559 0.5862206 0.5588695 0.5435777 0.5439967
#>  [29] 0.5477829 0.5532984 0.5465179 0.5790366 0.5477829 0.5452559 0.5414872
#>  [36] 0.5507459 0.5597315 0.5448358 0.5515955 0.5439967 0.5537250 0.5649318
#>  [43] 0.5524463 0.5498977 0.5575789 0.5503217 0.5688637 0.5482052 0.5554345
#>  [50] 0.5423224 0.5427405 0.5511705 0.5498977 0.5431589 0.5486279 0.5448358
#>  [57] 0.5490508 0.5537250 0.5439967 0.5456762 0.5494741 0.5498977 0.5465179
#>  [64] 0.5750344 0.5439967 0.5503217 0.5448358 0.5423224 0.5414872 0.5469393
#>  [71] 0.5477829 0.7484404 0.5435777 0.5452559 0.5431589 0.5439967 0.5414872
#>  [78] 0.5435777 0.5419046 0.5812721 0.5414872 0.5406532 0.5410700 0.5414872
#>  [85] 0.5414872 0.5410700 0.5679876 0.6385743 0.6739685 0.5406532 0.5423224
#>  [92] 0.5419046 0.5402367 0.5406532 0.5473609 0.5402367 0.5406532 0.5419046
#>  [99] 0.5427405 0.5419046 0.5431589 0.5406532 0.5571493 0.5439967 0.5439967
#> [106] 0.5444161 0.5406532 0.5419046 0.5419046 0.5414872 0.5745914 0.5427405
#> [113] 0.7229300 0.5431589 0.5398205 0.5427405 0.5419046 0.5853177 0.5473609
#> [120] 0.5431589 0.5439967 0.5410700 0.5675500 0.5414872 0.5477829 0.6560328
#> [127] 0.5419046 0.5452559 0.5444161 0.5465179 0.5490508 0.5684255 0.5541519
#> [134] 0.5666759 0.5812721 0.5866725 0.7594799