Chi-squared test for outliers

Performs a Chi-squared test for outliers.

outlier_chisq(
  x,
  mask = !is.na(x),
  threshold = c(0.9, 0.95),
  return.score = FALSE
)

Arguments

x
A vector of data.

mask
A logical vector that defines which values in x will used when computing statistics. Useful when a subset of quality-assured data is available. Default mask is non-NA Values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

Details

the values of threshold identify the quantiles of the Chi-squared distribution used to identify mild and extreme outliers. Default values are 0.9 for "mild" outliers and 0.95 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rnorm(length(x), mean = 1, sd = 3)
y = sin(x) + noise
mask = noise < 1

outlier_chisq(y)

#> Error: must be a character vector, not an integer vector.

outlier_chisq(y, mask)

#> Error: must be a character vector, not an integer vector.

outlier_chisq(y, mask, threshold = c(0.8, 0.9))

#> Error: must be a character vector, not an integer vector.

outlier_chisq(y, return.score = TRUE)

#>   [1] 2.772037e-02 1.266003e-02 9.719711e-01 2.962280e-02 2.929216e-02
#>   [6] 1.516655e+00 1.691936e-02 2.260291e-01 2.181057e+00 1.436629e-01
#>  [11] 1.221679e+00 4.000287e-01 9.042048e-02 1.825153e+00 7.587321e-02
#>  [16] 3.094239e-01 6.601371e-01 1.792849e+00 2.533262e+00 8.300513e-01
#>  [21] 7.724695e+00 8.707923e-02 4.389216e-01 1.135204e-01 6.427370e-01
#>  [26] 1.565078e+00 2.679336e-01 8.670671e-01 1.289895e+00 7.639112e-01
#>  [31] 1.824201e-01 1.574066e-01 4.304182e-01 7.980120e-01 5.202738e-02
#>  [36] 1.573758e-01 7.462782e+00 1.117468e+00 1.049742e+00 6.508581e-01
#>  [41] 5.370454e-01 2.991149e-02 5.041270e-01 6.567188e-01 6.911952e-01
#>  [46] 2.327749e-01 7.372058e-01 7.376153e-03 1.534823e+00 4.701280e+00
#>  [51] 1.244007e-01 2.953449e+00 2.830031e-01 1.262602e-01 8.556859e-01
#>  [56] 3.427147e-01 5.081392e-01 9.379375e-01 6.985486e-02 2.041057e-01
#>  [61] 7.405380e-01 2.794901e+00 1.363654e-02 1.552572e-02 5.070491e-01
#>  [66] 1.983590e-01 1.014561e-02 3.980512e+00 3.009587e-01 2.074022e-01
#>  [71] 1.497793e-01 3.469713e+00 2.385980e+00 5.427923e+00 3.651818e-01
#>  [76] 1.405766e-01 3.669347e-01 1.647105e-01 3.073677e-01 9.679123e-01
#>  [81] 1.602495e+00 2.362051e-01 4.109976e-02 1.350084e+00 1.652124e-01
#>  [86] 2.843449e+00 1.293230e+00 3.371901e-02 1.401343e-01 3.708147e-02
#>  [91] 1.164828e-01 4.746384e+00 7.255313e-02 4.700128e-02 1.466708e+00
#>  [96] 4.043434e-01 1.635389e+00 4.025037e-02 6.753176e-03 3.504161e+00
#> [101] 5.308320e-02 3.645603e-02 6.460663e-01 7.872417e-01 8.118525e-05
#> [106] 5.153059e-02 3.930438e-01 1.197362e+00 3.357932e+00 1.028017e-02
#> [111] 1.593801e-01 8.112612e-02 4.389889e+00 2.957677e-02 4.658043e-01
#> [116] 3.823007e-01 3.387800e-01 8.817434e-01 1.240015e-06 1.876993e+00
#> [121] 5.637383e+00 1.354041e+00 1.708942e-06 3.746648e-01 4.634978e-04
#> [126] 1.467629e+00 1.217127e+00 2.869035e-01 1.777134e+00 2.307861e+00
#> [131] 9.600995e-01 2.103118e+00 6.482555e-01 1.017618e+00 1.785937e+00
#> [136] 3.033443e-02 8.430916e-01