Isolation Forest outlier detection

Performs outlier detection using an Isolation Forest.

outlier_iforest(
  xs,
  mask = Reduce("|", lapply(xs, is.na)),
  threshold = c(0.8, 0.9),
  return.score = FALSE,
  ...
)

Arguments

xs
A dataframe or list of vectors (which will be coerced to a numeric matrix).

mask
A logical mask that identifies a subgroup of x to ignore when computing statistics. Useful when a subset of quality-assured data is available. Default action is to mask NA values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

...
Additional arguments to solitude::isolationForest$new(). note that the argument sample_size will be overwritten to use the number of unmasked data points, i.e. length(which(mask)).

Details

the values of threshold identify mild and extreme\ outliers based on the Isolation Forest score in the range [0,1]. Default values are 0.8 for "mild" outliers and 0.9 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y = sin(x) + noise
mask = noise > 1

outlier_iforest(list(y))

#> Loading required namespace: solitude

#> INFO  [22:25:09.479] Building Isolation Forest ...  
#> INFO  [22:25:10.671] done 
#> INFO  [22:25:10.682] Computing depth of terminal nodes ...  
#> INFO  [22:25:11.464] done 
#> INFO  [22:25:11.478] Completed growing isolation forest 

#>   [1] not outlier  not outlier  not outlier  not outlier  not outlier 
#>   [6] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [11] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [16] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [21] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [26] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [31] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [36] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [41] not outlier  mild outlier not outlier  not outlier  not outlier 
#>  [46] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [51] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [56] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [61] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [66] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [71] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [76] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [81] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [86] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [91] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [96] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [101] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [106] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [111] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [116] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [121] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [126] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [131] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [136] not outlier  not outlier 
#> Levels: not outlier < mild outlier < extreme outlier

outlier_iforest(list(x, y))

#> INFO  [22:25:11.505] Building Isolation Forest ...  
#> INFO  [22:25:11.514] done 
#> INFO  [22:25:11.516] Computing depth of terminal nodes ...  
#> INFO  [22:25:11.796] done 
#> INFO  [22:25:11.805] Completed growing isolation forest 

#>   [1] not outlier  not outlier  not outlier  not outlier  not outlier 
#>   [6] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [11] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [16] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [21] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [26] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [31] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [36] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [41] not outlier  mild outlier not outlier  not outlier  not outlier 
#>  [46] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [51] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [56] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [61] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [66] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [71] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [76] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [81] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [86] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [91] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [96] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [101] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [106] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [111] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [116] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [121] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [126] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [131] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [136] not outlier  not outlier 
#> Levels: not outlier < mild outlier < extreme outlier

outlier_iforest(list(x, y), mask)

#> INFO  [22:25:11.824] Building Isolation Forest ...  
#> INFO  [22:25:11.830] done 
#> INFO  [22:25:11.831] Computing depth of terminal nodes ...  
#> INFO  [22:25:12.075] done 
#> INFO  [22:25:12.083] Completed growing isolation forest 

#>   [1] not outlier not outlier not outlier not outlier not outlier not outlier
#>   [7] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [13] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [19] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [25] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [31] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [37] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [43] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [49] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [55] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [61] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [67] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [73] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [79] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [85] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [91] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [97] not outlier not outlier not outlier not outlier not outlier not outlier
#> [103] not outlier not outlier not outlier not outlier not outlier not outlier
#> [109] not outlier not outlier not outlier not outlier not outlier not outlier
#> [115] not outlier not outlier not outlier not outlier not outlier not outlier
#> [121] not outlier not outlier not outlier not outlier not outlier not outlier
#> [127] not outlier not outlier not outlier not outlier not outlier not outlier
#> [133] not outlier not outlier not outlier not outlier not outlier
#> Levels: not outlier < mild outlier < extreme outlier

outlier_iforest(list(x, y), mask, threshold = c(1, 2))

#> INFO  [22:25:12.101] Building Isolation Forest ...  
#> INFO  [22:25:12.108] done 
#> INFO  [22:25:12.110] Computing depth of terminal nodes ...  
#> INFO  [22:25:12.379] done 
#> INFO  [22:25:12.387] Completed growing isolation forest 

#>   [1] not outlier not outlier not outlier not outlier not outlier not outlier
#>   [7] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [13] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [19] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [25] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [31] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [37] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [43] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [49] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [55] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [61] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [67] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [73] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [79] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [85] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [91] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [97] not outlier not outlier not outlier not outlier not outlier not outlier
#> [103] not outlier not outlier not outlier not outlier not outlier not outlier
#> [109] not outlier not outlier not outlier not outlier not outlier not outlier
#> [115] not outlier not outlier not outlier not outlier not outlier not outlier
#> [121] not outlier not outlier not outlier not outlier not outlier not outlier
#> [127] not outlier not outlier not outlier not outlier not outlier not outlier
#> [133] not outlier not outlier not outlier not outlier not outlier
#> Levels: not outlier < mild outlier < extreme outlier

outlier_iforest(list(x, y), return.score = TRUE)

#> INFO  [22:25:12.412] Building Isolation Forest ...  
#> INFO  [22:25:12.419] done 
#> INFO  [22:25:12.421] Computing depth of terminal nodes ...  
#> INFO  [22:25:12.654] done 
#> INFO  [22:25:12.672] Completed growing isolation forest 

#>   [1] 0.5631930 0.6570447 0.5601630 0.5567201 0.5537250 0.5723817 0.5875774
#>   [8] 0.5640617 0.5541519 0.5486279 0.5520207 0.5701804 0.5759214 0.5469393
#>  [15] 0.5821687 0.5541519 0.5857690 0.5507459 0.5545791 0.5469393 0.5524463
#>  [22] 0.5623256 0.5503217 0.5452559 0.5515955 0.5482052 0.5482052 0.5584389
#>  [29] 0.5477829 0.5448358 0.5507459 0.5524463 0.5477829 0.5912112 0.5482052
#>  [36] 0.5532984 0.5444161 0.5465179 0.5486279 0.5444161 0.5541519 0.8664589
#>  [43] 0.5444161 0.5507459 0.5473609 0.5456762 0.5419046 0.5482052 0.5439967
#>  [50] 0.5862206 0.5419046 0.5423224 0.5414872 0.5423224 0.5423224 0.5427405
#>  [57] 0.5593003 0.6405458 0.6823303 0.5419046 0.5423224 0.5406532 0.5423224
#>  [64] 0.5427405 0.5473609 0.5423224 0.5423224 0.5410700 0.5410700 0.5431589
#>  [71] 0.5431589 0.5414872 0.5511705 0.5402367 0.5419046 0.5398205 0.5406532
#>  [78] 0.5398205 0.5398205 0.5406532 0.5545791 0.5398205 0.7257210 0.5427405
#>  [85] 0.5402367 0.5410700 0.5410700 0.5584389 0.5402367 0.5402367 0.5402367
#>  [92] 0.5402367 0.5477829 0.5402367 0.5406532 0.5934936 0.5419046 0.5406532
#>  [99] 0.5427405 0.5414872 0.5410700 0.5423224 0.5402367 0.5419046 0.5435777
#> [106] 0.5448358 0.6450040 0.5423224 0.5507459 0.5419046 0.5427405 0.5406532
#> [113] 0.5419046 0.5410700 0.5423224 0.6361185 0.5423224 0.5423224 0.5431589
#> [120] 0.5456762 0.5494741 0.5410700 0.5402367 0.5402367 0.5653673 0.5465179
#> [127] 0.5550066 0.5419046 0.5439967 0.5490508 0.5545791 0.5554345 0.5567201
#> [134] 0.6059721 0.5662394 0.5754777 0.5948673