Median absolute deviation test for outliers

Performs a median absolute deviation (MAD) test for outliers.

outlier_mad(
  x,
  mask = !is.na(x),
  threshold = c(1.5, 3),
  k = 1/qnorm(0.75),
  return.score = FALSE
)

Arguments

x
A vector of data.

mask
A logical vector that defines which values in x will used when computing statistics. Useful when a subset of quality-assured data is available. Default mask is non-NA Values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

k
A scale factor, defined as the reciprocal of the 75th percentile of the underlying distribution of the data. Default assumes a normal distribution.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

Details

the values of threshold identify the thresholds used to identify mild and extreme outliers, as a multiple of k * median(x). Default values are 1.5 for "mild" outliers and 3.0 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rnorm(length(x), mean = 1, sd = 3)
y = sin(x) + noise
mask = noise < 1

outlier_mad(y)

#>   [1] not outlier  not outlier  mild outlier not outlier  not outlier 
#>   [6] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [11] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [16] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [21] not outlier  not outlier  mild outlier not outlier  not outlier 
#>  [26] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [31] not outlier  not outlier  not outlier  mild outlier not outlier 
#>  [36] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [41] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [46] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [51] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [56] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [61] not outlier  not outlier  mild outlier mild outlier mild outlier
#>  [66] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [71] not outlier  not outlier  not outlier  not outlier  mild outlier
#>  [76] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [81] not outlier  not outlier  not outlier  mild outlier not outlier 
#>  [86] not outlier  not outlier  not outlier  mild outlier not outlier 
#>  [91] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [96] mild outlier mild outlier mild outlier not outlier  not outlier 
#> [101] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [106] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [111] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [116] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [121] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [126] not outlier  not outlier  not outlier  not outlier  mild outlier
#> [131] mild outlier not outlier  not outlier  not outlier  not outlier 
#> [136] mild outlier not outlier 
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

outlier_mad(y, mask)

#>   [1] not outlier     not outlier     mild outlier    not outlier    
#>   [5] not outlier     not outlier     not outlier     not outlier    
#>   [9] not outlier     not outlier     not outlier     not outlier    
#>  [13] not outlier     not outlier     not outlier     not outlier    
#>  [17] mild outlier    not outlier     not outlier     not outlier    
#>  [21] mild outlier    not outlier     mild outlier    not outlier    
#>  [25] not outlier     not outlier     not outlier     mild outlier   
#>  [29] not outlier     not outlier     not outlier     not outlier    
#>  [33] not outlier     mild outlier    not outlier     not outlier    
#>  [37] not outlier     not outlier     not outlier     not outlier    
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     not outlier    
#>  [49] not outlier     not outlier     not outlier     not outlier    
#>  [53] not outlier     not outlier     not outlier     not outlier    
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     not outlier     mild outlier    not outlier    
#>  [65] mild outlier    not outlier     not outlier     not outlier    
#>  [69] not outlier     not outlier     not outlier     not outlier    
#>  [73] mild outlier    not outlier     extreme outlier not outlier    
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     not outlier     not outlier     mild outlier   
#>  [85] mild outlier    not outlier     not outlier     not outlier    
#>  [89] mild outlier    not outlier     not outlier     not outlier    
#>  [93] not outlier     not outlier     not outlier     mild outlier   
#>  [97] mild outlier    mild outlier    not outlier     not outlier    
#> [101] not outlier     not outlier     not outlier     not outlier    
#> [105] not outlier     not outlier     not outlier     not outlier    
#> [109] mild outlier    not outlier     not outlier     not outlier    
#> [113] not outlier     mild outlier    not outlier     not outlier    
#> [117] not outlier     not outlier     not outlier     not outlier    
#> [121] not outlier     not outlier     not outlier     not outlier    
#> [125] not outlier     not outlier     not outlier     not outlier    
#> [129] not outlier     mild outlier    mild outlier    not outlier    
#> [133] not outlier     not outlier     not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

outlier_mad(y, mask, threshold = c(1, 2))

#>   [1] not outlier     mild outlier    mild outlier    mild outlier   
#>   [5] not outlier     not outlier     not outlier     not outlier    
#>   [9] not outlier     not outlier     not outlier     not outlier    
#>  [13] not outlier     mild outlier    mild outlier    not outlier    
#>  [17] mild outlier    not outlier     mild outlier    not outlier    
#>  [21] mild outlier    mild outlier    extreme outlier not outlier    
#>  [25] not outlier     not outlier     mild outlier    mild outlier   
#>  [29] not outlier     not outlier     not outlier     not outlier    
#>  [33] not outlier     extreme outlier not outlier     mild outlier   
#>  [37] not outlier     not outlier     not outlier     mild outlier   
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     not outlier    
#>  [49] not outlier     not outlier     mild outlier    not outlier    
#>  [53] mild outlier    mild outlier    not outlier     not outlier    
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     not outlier     extreme outlier mild outlier   
#>  [65] extreme outlier mild outlier    mild outlier    not outlier    
#>  [69] not outlier     not outlier     mild outlier    not outlier    
#>  [73] mild outlier    not outlier     extreme outlier mild outlier   
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     not outlier     mild outlier    extreme outlier
#>  [85] mild outlier    not outlier     mild outlier    not outlier    
#>  [89] extreme outlier not outlier     not outlier     not outlier    
#>  [93] not outlier     not outlier     not outlier     extreme outlier
#>  [97] extreme outlier extreme outlier not outlier     not outlier    
#> [101] not outlier     not outlier     not outlier     mild outlier   
#> [105] not outlier     not outlier     mild outlier    mild outlier   
#> [109] mild outlier    not outlier     not outlier     not outlier    
#> [113] not outlier     mild outlier    not outlier     not outlier    
#> [117] mild outlier    not outlier     not outlier     not outlier    
#> [121] not outlier     not outlier     mild outlier    not outlier    
#> [125] mild outlier    mild outlier    not outlier     not outlier    
#> [129] not outlier     extreme outlier extreme outlier not outlier    
#> [133] not outlier     not outlier     not outlier     mild outlier   
#> [137] not outlier    
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

outlier_mad(y, return.score = TRUE)

#>   [1] 0.48660436 0.95959219 2.11455075 0.70311726 0.05808117 0.19233627
#>   [7] 0.41862324 0.11669323 1.16130841 0.23508626 0.92349837 0.42110042
#>  [13] 0.04803104 0.87325170 0.87123645 0.63497684 1.29912406 0.67448975
#>  [19] 0.93569094 0.56003855 1.25130821 0.70824592 2.35852982 0.77336909
#>  [25] 0.55176209 0.53304532 0.90677976 1.13782520 0.17514866 0.44676280
#>  [31] 0.46741069 0.17695575 1.38867042 2.07563146 0.46886421 1.08113111
#>  [37] 0.01958725 0.22854654 1.17268947 0.99634464 0.66151754 0.39839916
#>  [43] 0.24615635 0.35111027 0.24874417 0.94158468 0.30322537 1.48281401
#>  [49] 0.46802018 0.45553034 0.81335752 0.02668393 0.70935225 0.82572464
#>  [55] 0.08489289 0.16607785 0.47052074 0.42149926 0.08097559 1.25035126
#>  [61] 0.32705127 0.57824219 2.33765520 1.79585292 2.94396478 1.05511418
#>  [67] 1.05323215 0.17671474 0.70748493 1.08912200 0.91626379 0.46894481
#>  [73] 1.19941824 0.20448137 2.92441825 0.77184510 0.11366365 0.58482543
#>  [79] 0.81086575 0.23346463 0.39105006 0.12857806 0.75738582 1.77073635
#>  [85] 1.27377954 0.38960210 0.65211632 0.13945375 2.00327315 1.16790661
#>  [91] 0.99048645 0.09610228 1.33437090 0.42522433 0.40505492 1.67275284
#>  [97] 1.79544187 1.71454689 0.29943706 1.00368642 0.26837859 1.19930366
#> [103] 1.04459701 0.87159624 0.16746323 0.04371186 0.64558719 0.72545893
#> [109] 1.27966625 0.00000000 0.14747310 1.08117778 0.58585020 1.18310145
#> [115] 0.20607501 0.58912416 0.82451120 0.60021647 0.39428180 1.02123886
#> [121] 0.59274670 0.58433886 0.89783594 0.88849739 0.92862719 0.96905272
#> [127] 0.11122828 1.35538016 0.55164562 1.80334128 2.06451713 0.68918898
#> [133] 0.30450476 0.29318336 0.33727738 1.73662779 1.02779015