t-score test for outliers

Performs a t-score test for outliers.

outlier_tscore(
  x,
  mask = !is.na(x),
  threshold = c(0.9, 0.95),
  return.score = FALSE
)

Arguments

x
A vector of data.

mask
A logical mask that identifies a subgroup of x to compute statistics on. Useful when a subset of quality-assured data is available. Default action is to ignore NA values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

Details

The values of threshold identify the quantiles of the t-distribution used to identify mild and extreme outliers. Default values are 0.9 for "mild" outliers and 0.95 for "extreme" outliers.

The t-score is equivalent to the z-score for sample sizes greater than 30.

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

outlier_tscore(y)

#>   [1] extreme outlier not outlier     extreme outlier extreme outlier
#>   [5] extreme outlier extreme outlier mild outlier    extreme outlier
#>   [9] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [13] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [17] extreme outlier mild outlier    not outlier     extreme outlier
#>  [21] extreme outlier extreme outlier mild outlier    extreme outlier
#>  [25] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [29] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [33] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [37] extreme outlier mild outlier    extreme outlier extreme outlier
#>  [41] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [45] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [49] extreme outlier not outlier     mild outlier    extreme outlier
#>  [53] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [57] not outlier     extreme outlier extreme outlier extreme outlier
#>  [61] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [65] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [69] extreme outlier not outlier     not outlier     extreme outlier
#>  [73] extreme outlier extreme outlier not outlier     extreme outlier
#>  [77] extreme outlier extreme outlier not outlier     extreme outlier
#>  [81] mild outlier    extreme outlier extreme outlier extreme outlier
#>  [85] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [89] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [93] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [97] extreme outlier extreme outlier extreme outlier extreme outlier
#> [101] extreme outlier extreme outlier extreme outlier extreme outlier
#> [105] mild outlier    extreme outlier extreme outlier extreme outlier
#> [109] extreme outlier mild outlier    mild outlier    extreme outlier
#> [113] extreme outlier extreme outlier extreme outlier not outlier    
#> [117] extreme outlier not outlier     not outlier     extreme outlier
#> [121] extreme outlier extreme outlier extreme outlier extreme outlier
#> [125] extreme outlier extreme outlier extreme outlier extreme outlier
#> [129] not outlier     mild outlier    extreme outlier not outlier    
#> [133] extreme outlier extreme outlier extreme outlier extreme outlier
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

outlier_tscore(y, mask)

#>   [1] extreme outlier extreme outlier extreme outlier extreme outlier
#>   [5] mild outlier    extreme outlier extreme outlier extreme outlier
#>   [9] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [13] not outlier     extreme outlier extreme outlier extreme outlier
#>  [17] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [21] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [25] mild outlier    extreme outlier extreme outlier extreme outlier
#>  [29] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [33] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [37] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [41] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [45] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [49] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [53] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [57] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [61] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [65] not outlier     extreme outlier extreme outlier extreme outlier
#>  [69] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [73] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [77] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [81] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [85] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [89] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [93] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [97] extreme outlier extreme outlier extreme outlier extreme outlier
#> [101] extreme outlier extreme outlier extreme outlier extreme outlier
#> [105] extreme outlier extreme outlier extreme outlier extreme outlier
#> [109] extreme outlier extreme outlier extreme outlier extreme outlier
#> [113] extreme outlier extreme outlier extreme outlier extreme outlier
#> [117] extreme outlier extreme outlier extreme outlier extreme outlier
#> [121] not outlier     extreme outlier extreme outlier extreme outlier
#> [125] extreme outlier extreme outlier extreme outlier extreme outlier
#> [129] extreme outlier extreme outlier extreme outlier extreme outlier
#> [133] extreme outlier extreme outlier extreme outlier extreme outlier
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

outlier_tscore(y, mask, threshold = c(0.8, 0.9))

#>   [1] extreme outlier extreme outlier extreme outlier extreme outlier
#>   [5] extreme outlier extreme outlier extreme outlier extreme outlier
#>   [9] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [13] mild outlier    extreme outlier extreme outlier extreme outlier
#>  [17] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [21] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [25] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [29] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [33] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [37] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [41] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [45] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [49] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [53] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [57] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [61] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [65] not outlier     extreme outlier extreme outlier extreme outlier
#>  [69] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [73] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [77] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [81] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [85] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [89] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [93] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [97] extreme outlier extreme outlier extreme outlier extreme outlier
#> [101] extreme outlier extreme outlier extreme outlier extreme outlier
#> [105] extreme outlier extreme outlier extreme outlier extreme outlier
#> [109] extreme outlier extreme outlier extreme outlier extreme outlier
#> [113] extreme outlier extreme outlier extreme outlier extreme outlier
#> [117] extreme outlier extreme outlier extreme outlier extreme outlier
#> [121] not outlier     extreme outlier extreme outlier extreme outlier
#> [125] extreme outlier extreme outlier extreme outlier extreme outlier
#> [129] extreme outlier extreme outlier extreme outlier extreme outlier
#> [133] extreme outlier extreme outlier extreme outlier extreme outlier
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

outlier_tscore(y, return.score = TRUE)

#>   [1]  -1.86280857  -0.71879648  -1.91389844  -1.92328934  -1.95956937
#>   [6]  -1.84253381  -1.52788580  -1.96429021  -1.89823762  -1.67853051
#>  [11]  -1.94555466  -1.77276705  -1.95934265  -1.96510378  -1.96929069
#>  [16]  -1.83879992  -1.96013012  -1.56579582  -1.05200733  -1.91944555
#>  [21]  -1.96018659  -1.96473176   1.45625033   4.08500855  -1.95968737
#>  [26]  -1.94691690  -1.93666462  -1.93737169  -1.96048686  -1.96705555
#>  [31]  -1.96987452  -1.95130010  -1.94239143  -1.94722485  -1.94354696
#>  [36]  -1.95349302  -1.95398695  -1.60588223  -1.97053260   7.59171755
#>  [41]  -1.93855224  -1.76339792  -1.95049995  -1.88425556  -1.95161134
#>  [46]  -1.96517211  -1.97057712  -1.94432513  -1.89846352  -0.42937116
#>  [51]  -1.33438228  -1.95325722  -1.96028898  -1.96763010  -1.91146001
#>  [56]  -1.43833146  -0.02845006  -1.95313669  -1.93069735  -1.93283050
#>  [61]  -1.95486562  -1.96591797  -1.97021510  -1.96629371  -1.95822102
#>  [66]  -1.95322150  -1.67033336  -1.62992323  -1.90234014  -0.18393128
#>  [71]   0.07089264  -1.95968067  -1.85148330  -1.95090704  -1.23854948
#>  [76]  -1.92396203  -1.87175807  -1.95468463  -1.26809295  -1.91144824
#>  [81]  -1.42719890  -1.93854447  -1.88797594  -1.95046139  -1.96014823
#>  [86]   5.26253205  -1.85373343  -1.70360023  -1.90718224  -1.94231360
#>  [91]  -1.85264707  -1.87852917  -1.87941466   2.37916655  -1.96812757
#>  [96]  -1.94441887  -1.94869550  -1.95271878  -1.94780738  -1.95043181
#> [101]  -1.96052224  -1.96313554  -1.97003606  -1.85492948  -1.55352259
#> [106]  -1.93366414  -1.88148504  -1.87917930  -1.92099520  -1.60548034
#> [111]  -1.65198665  -1.95228523  -1.95041271  -1.94741427  -1.95034569
#> [116]  -0.85273107  -1.93132777   0.87271413  -1.08326582  -1.96756267
#> [121]  -1.95868794  45.52756921  -1.88365421  -1.90978693  -1.95561023
#> [126] 121.59583783  -1.95371653  -1.94275583   0.81932061  -1.63403566
#> [131]  -1.95008231   1.01565197  -1.87454785  -1.96647189  -1.91153652
#> [136]  -1.93043155  35.20081886