t-score test for outliers

Performs a t-score test for outliers.

tscore_outliers(
  x,
  mask = !is.na(x),
  threshold = c(0.9, 0.95),
  return.score = FALSE
)

Arguments

x
A vector of data.

mask
A logical mask that identifies a subgroup of x to compute non-parametric statistics on. Useful when a subset of quality-assured data is available. Default action is to ignore NA values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

Details

The values of threshold identify the quantiles of the t-distribution used to identify mild and extreme outliers. Default values are 0.9 for "mild" outliers and 0.95 for "extreme" outliers.

The t-score is equivalent to the z-score for sample sizes greater than 30.

Examples

x = seq(0, 34, by = 0.25)*pi     
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

tscore_outliers(y)

#>   [1] extreme outlier extreme outlier extreme outlier extreme outlier
#>   [5] extreme outlier extreme outlier not outlier     extreme outlier
#>   [9] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [13] extreme outlier not outlier     extreme outlier extreme outlier
#>  [17] extreme outlier extreme outlier mild outlier    extreme outlier
#>  [21] extreme outlier not outlier     extreme outlier extreme outlier
#>  [25] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [29] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [33] extreme outlier extreme outlier mild outlier    extreme outlier
#>  [37] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [41] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [45] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [49] extreme outlier extreme outlier mild outlier    extreme outlier
#>  [53] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [57] extreme outlier extreme outlier extreme outlier not outlier    
#>  [61] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [65] extreme outlier mild outlier    extreme outlier extreme outlier
#>  [69] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [73] extreme outlier mild outlier    extreme outlier extreme outlier
#>  [77] extreme outlier mild outlier    extreme outlier extreme outlier
#>  [81] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [85] extreme outlier extreme outlier extreme outlier not outlier    
#>  [89] not outlier     not outlier     extreme outlier not outlier    
#>  [93] not outlier     extreme outlier extreme outlier extreme outlier
#>  [97] not outlier     extreme outlier mild outlier    extreme outlier
#> [101] extreme outlier extreme outlier extreme outlier extreme outlier
#> [105] extreme outlier extreme outlier extreme outlier extreme outlier
#> [109] extreme outlier extreme outlier not outlier     extreme outlier
#> [113] mild outlier    extreme outlier extreme outlier extreme outlier
#> [117] extreme outlier extreme outlier extreme outlier not outlier    
#> [121] not outlier     extreme outlier extreme outlier extreme outlier
#> [125] not outlier     extreme outlier extreme outlier extreme outlier
#> [129] extreme outlier extreme outlier extreme outlier extreme outlier
#> [133] extreme outlier extreme outlier extreme outlier extreme outlier
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

tscore_outliers(y, mask)

#>   [1] extreme outlier extreme outlier extreme outlier extreme outlier
#>   [5] not outlier     extreme outlier extreme outlier extreme outlier
#>   [9] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [13] extreme outlier extreme outlier extreme outlier not outlier    
#>  [17] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [21] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [25] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [29] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [33] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [37] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [41] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [45] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [49] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [53] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [57] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [61] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [65] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [69] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [73] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [77] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [81] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [85] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [89] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [93] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [97] extreme outlier extreme outlier extreme outlier extreme outlier
#> [101] extreme outlier extreme outlier extreme outlier extreme outlier
#> [105] extreme outlier extreme outlier extreme outlier extreme outlier
#> [109] extreme outlier extreme outlier extreme outlier extreme outlier
#> [113] extreme outlier extreme outlier extreme outlier extreme outlier
#> [117] extreme outlier extreme outlier extreme outlier extreme outlier
#> [121] extreme outlier extreme outlier extreme outlier extreme outlier
#> [125] extreme outlier extreme outlier extreme outlier extreme outlier
#> [129] extreme outlier extreme outlier extreme outlier extreme outlier
#> [133] extreme outlier extreme outlier extreme outlier extreme outlier
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

tscore_outliers(y, mask, threshold = c(0.8, 0.9))

#>   [1] extreme outlier extreme outlier extreme outlier extreme outlier
#>   [5] mild outlier    extreme outlier extreme outlier extreme outlier
#>   [9] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [13] extreme outlier extreme outlier extreme outlier not outlier    
#>  [17] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [21] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [25] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [29] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [33] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [37] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [41] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [45] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [49] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [53] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [57] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [61] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [65] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [69] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [73] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [77] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [81] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [85] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [89] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [93] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [97] extreme outlier extreme outlier extreme outlier extreme outlier
#> [101] extreme outlier extreme outlier extreme outlier extreme outlier
#> [105] extreme outlier extreme outlier extreme outlier extreme outlier
#> [109] extreme outlier extreme outlier extreme outlier extreme outlier
#> [113] extreme outlier extreme outlier extreme outlier extreme outlier
#> [117] extreme outlier extreme outlier extreme outlier extreme outlier
#> [121] extreme outlier extreme outlier extreme outlier extreme outlier
#> [125] extreme outlier extreme outlier extreme outlier extreme outlier
#> [129] extreme outlier extreme outlier extreme outlier extreme outlier
#> [133] extreme outlier extreme outlier extreme outlier extreme outlier
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

tscore_outliers(y, return.score = TRUE)

#>   [1]  -1.90996478  -2.09407011  -2.11911200  -2.12549693  -2.13519531
#>   [6]  -2.09890322  -0.87455058  -2.00705067  32.15354575  -2.07463223
#>  [11]  -2.12687642  -2.12039225  -2.12871361  -0.69050462  -2.15417941
#>  [16]  -2.13573383  -2.10547049  -2.13097948  -1.58594414  -2.13056988
#>  [21]  -1.99800481   0.13238197  -2.11925967  -2.12349860  -2.05784031
#>  [26]  -2.03774609  -2.10534774  -1.93010589  -2.11934108  -2.14723150
#>  [31]  -2.15427446  -1.97848424   5.09565607  -1.90936466  -1.33540608
#>  [36]  -2.07702191  -2.13928210  -2.14929578  -2.14689579  -2.14565367
#>  [41]  -2.12033833   3.73750980  -2.12528328  -2.10753320  -2.10814698
#>  [46]  -2.14836421  -2.02388050  -2.09773802  -2.09535339  -2.12863122
#>  [51]  -1.56700356  -2.12758473  -1.86250472  -2.09079806  -2.09940432
#>  [56]  -2.15013181  -2.07197589  -2.12665361  -2.01603348   0.36109703
#>  [61]  -2.02609906  -2.15016692  -2.12244980  -1.97527587  -1.94564755
#>  [66]  -1.62461127  -1.83491216  -2.10800057  -2.14058117  -2.14883953
#>  [71]  -1.70226216  -1.91870049  -2.11477900  -1.51500909  -2.12487358
#>  [76]  -2.05043919  -2.13346011  -1.53959715  -2.14707740  -2.14086241
#>  [81]  -2.13072441  -2.12510684  -1.98596762  -1.97156911  -1.95013449
#>  [86]  -2.15047004  -2.05944788   0.37982668   0.19297155  -1.22860890
#>  [91]  -1.87723745  -1.15968985   0.05273015  -2.03932978  22.97761446
#>  [96]  -2.09361078   0.18743938  -2.12242756   1.61968103  -2.12025434
#> [101]  -2.13289503  -1.66962690  -2.05230379  -2.14739517  -2.12130985
#> [106]  -2.13064459  -2.05151285  -2.08986426  -2.03108954   2.52336222
#> [111]   0.57921194  -2.14832991  -1.35158453  -2.08903331  -2.12676948
#> [116]  -2.12896688  -2.12190340  -2.14849646 118.53629407   0.29266158
#> [121]   0.07665938  -2.11848460  -2.12420455  -2.12280776  -0.70708958
#> [126]  -1.97898091  -2.15447770  -2.10899167  -1.65677240  -1.78253742
#> [131]  -1.92756597  -2.13012305  -2.12066808  -2.01678479  -2.14635050
#> [136]  -2.14452361  49.85142262