Tukey's test for outliers

Perform Tukey's test for mild and extreme outliers.

tukey_outliers(
  x,
  mask = !is.na(x),
  threshold = c(1.5, 3),
  return.score = FALSE
)

Arguments

x
A vector of data.

mask
A logical mask that identifies a subgroup of x to compute non-parametric statistics on. Useful when a subset of quality-assured data is available. Default action is to ignore NA values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

Value

A vector the same length as x containing numeric scores or ordered factors.

Details

the values of threshold identify the multiplier of the interquartile range used to identify mild and extreme outliers. typical values are 1.5 for "mild" outliers and 3.0 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi     
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

tukey_outliers(y)

#>   [1] not outlier     not outlier     not outlier     not outlier    
#>   [5] not outlier     not outlier     not outlier     not outlier    
#>   [9] mild outlier    mild outlier    not outlier     not outlier    
#>  [13] not outlier     not outlier     not outlier     not outlier    
#>  [17] not outlier     not outlier     extreme outlier not outlier    
#>  [21] not outlier     not outlier     not outlier     not outlier    
#>  [25] not outlier     not outlier     not outlier     not outlier    
#>  [29] not outlier     not outlier     extreme outlier not outlier    
#>  [33] not outlier     not outlier     not outlier     not outlier    
#>  [37] not outlier     not outlier     not outlier     not outlier    
#>  [41] not outlier     extreme outlier not outlier     not outlier    
#>  [45] extreme outlier not outlier     not outlier     not outlier    
#>  [49] not outlier     not outlier     not outlier     not outlier    
#>  [53] mild outlier    not outlier     mild outlier    not outlier    
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     extreme outlier extreme outlier extreme outlier
#>  [65] not outlier     not outlier     not outlier     not outlier    
#>  [69] extreme outlier extreme outlier not outlier     not outlier    
#>  [73] not outlier     not outlier     not outlier     not outlier    
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     not outlier     extreme outlier extreme outlier
#>  [85] extreme outlier not outlier     not outlier     not outlier    
#>  [89] not outlier     not outlier     not outlier     not outlier    
#>  [93] not outlier     not outlier     not outlier     extreme outlier
#>  [97] mild outlier    not outlier     not outlier     mild outlier   
#> [101] not outlier     not outlier     not outlier     not outlier    
#> [105] not outlier     not outlier     not outlier     extreme outlier
#> [109] not outlier     mild outlier    not outlier     not outlier    
#> [113] not outlier     extreme outlier extreme outlier not outlier    
#> [117] not outlier     extreme outlier not outlier     not outlier    
#> [121] not outlier     not outlier     not outlier     extreme outlier
#> [125] not outlier     not outlier     not outlier     not outlier    
#> [129] not outlier     extreme outlier not outlier     not outlier    
#> [133] not outlier     not outlier     not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < mild outlier < extreme outlier

tukey_outliers(y, mask)

#>   [1] not outlier     not outlier     extreme outlier not outlier    
#>   [5] not outlier     extreme outlier extreme outlier not outlier    
#>   [9] extreme outlier extreme outlier not outlier     extreme outlier
#>  [13] extreme outlier not outlier     not outlier     not outlier    
#>  [17] not outlier     not outlier     extreme outlier not outlier    
#>  [21] extreme outlier not outlier     not outlier     not outlier    
#>  [25] not outlier     not outlier     mild outlier    not outlier    
#>  [29] not outlier     mild outlier    extreme outlier extreme outlier
#>  [33] extreme outlier not outlier     not outlier     not outlier    
#>  [37] not outlier     mild outlier    not outlier     not outlier    
#>  [41] not outlier     extreme outlier extreme outlier extreme outlier
#>  [45] extreme outlier not outlier     extreme outlier not outlier    
#>  [49] extreme outlier extreme outlier not outlier     mild outlier   
#>  [53] extreme outlier not outlier     extreme outlier not outlier    
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     extreme outlier extreme outlier extreme outlier
#>  [65] not outlier     not outlier     extreme outlier not outlier    
#>  [69] extreme outlier extreme outlier extreme outlier not outlier    
#>  [73] extreme outlier extreme outlier not outlier     extreme outlier
#>  [77] mild outlier    mild outlier    not outlier     not outlier    
#>  [81] mild outlier    not outlier     extreme outlier extreme outlier
#>  [85] extreme outlier not outlier     not outlier     not outlier    
#>  [89] not outlier     not outlier     not outlier     extreme outlier
#>  [93] extreme outlier mild outlier    not outlier     extreme outlier
#>  [97] extreme outlier not outlier     extreme outlier extreme outlier
#> [101] extreme outlier not outlier     not outlier     not outlier    
#> [105] not outlier     extreme outlier extreme outlier extreme outlier
#> [109] extreme outlier extreme outlier not outlier     not outlier    
#> [113] extreme outlier extreme outlier extreme outlier extreme outlier
#> [117] extreme outlier extreme outlier not outlier     extreme outlier
#> [121] extreme outlier not outlier     mild outlier    extreme outlier
#> [125] extreme outlier not outlier     not outlier     not outlier    
#> [129] mild outlier    extreme outlier extreme outlier not outlier    
#> [133] extreme outlier not outlier     extreme outlier extreme outlier
#> [137] not outlier    
#> Levels: not outlier < mild outlier < extreme outlier

tukey_outliers(y, mask, threshold = c(2, 5))

#>   [1] not outlier     not outlier     extreme outlier not outlier    
#>   [5] not outlier     extreme outlier extreme outlier not outlier    
#>   [9] extreme outlier extreme outlier not outlier     extreme outlier
#>  [13] extreme outlier not outlier     not outlier     not outlier    
#>  [17] not outlier     not outlier     extreme outlier not outlier    
#>  [21] extreme outlier not outlier     not outlier     not outlier    
#>  [25] not outlier     not outlier     mild outlier    not outlier    
#>  [29] not outlier     not outlier     extreme outlier extreme outlier
#>  [33] extreme outlier not outlier     not outlier     not outlier    
#>  [37] not outlier     mild outlier    not outlier     not outlier    
#>  [41] not outlier     extreme outlier mild outlier    extreme outlier
#>  [45] extreme outlier not outlier     extreme outlier not outlier    
#>  [49] extreme outlier mild outlier    not outlier     mild outlier   
#>  [53] extreme outlier not outlier     extreme outlier not outlier    
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     extreme outlier extreme outlier extreme outlier
#>  [65] not outlier     not outlier     extreme outlier not outlier    
#>  [69] extreme outlier extreme outlier extreme outlier not outlier    
#>  [73] extreme outlier extreme outlier not outlier     extreme outlier
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     not outlier     extreme outlier extreme outlier
#>  [85] extreme outlier not outlier     not outlier     not outlier    
#>  [89] not outlier     not outlier     not outlier     extreme outlier
#>  [93] extreme outlier not outlier     not outlier     extreme outlier
#>  [97] extreme outlier not outlier     mild outlier    extreme outlier
#> [101] extreme outlier not outlier     not outlier     not outlier    
#> [105] not outlier     mild outlier    extreme outlier extreme outlier
#> [109] mild outlier    extreme outlier not outlier     not outlier    
#> [113] extreme outlier extreme outlier extreme outlier extreme outlier
#> [117] extreme outlier extreme outlier not outlier     mild outlier   
#> [121] mild outlier    not outlier     mild outlier    extreme outlier
#> [125] mild outlier    not outlier     not outlier     not outlier    
#> [129] mild outlier    extreme outlier extreme outlier not outlier    
#> [133] extreme outlier not outlier     extreme outlier extreme outlier
#> [137] not outlier    
#> Levels: not outlier < mild outlier < extreme outlier

tukey_outliers(y, return.score = TRUE)

#>   [1] -7.080369e-02 -5.389600e-02  0.000000e+00  0.000000e+00  5.329147e-02
#>   [6] -1.306583e-01  9.791176e-01  8.911996e-02  2.087689e+00  2.520271e+00
#>  [11] -2.496208e-02 -4.090994e-01 -5.056810e-02  2.658088e-02  3.075714e-02
#>  [16]  2.183894e-02  4.689459e-02 -6.392310e-03  9.597920e+01 -1.270449e-02
#>  [21] -2.451861e-01 -7.074368e-02  1.025246e-01  1.046547e-01 -9.145862e-03
#>  [26] -4.263374e-04 -2.034170e-01 -1.805692e-02  4.072734e-02 -1.476794e-01
#>  [31]  7.260412e+00  6.569795e-01 -2.248461e-02 -7.822429e-03 -6.174612e-02
#>  [36] -2.326985e-02 -6.569131e-02 -2.283306e-01  1.059726e-01  6.628373e-02
#>  [41]  1.999957e-02  6.089622e+00 -3.820101e-01 -4.307642e-01  3.668920e+00
#>  [46]  1.027301e-01 -4.991295e-01  6.854920e-02 -4.526600e-01 -3.975766e-01
#>  [51] -3.278071e-02 -2.383165e-01  1.988376e+00  4.448147e-02  2.449307e+00
#>  [56]  7.199526e-02  5.138901e-02 -2.629442e-03 -2.498772e-02 -2.178038e-02
#>  [61]  3.932258e-03  1.175514e+01  2.997984e+01  5.873912e+00 -7.695523e-02
#>  [66] -1.192630e-01  1.296028e+00 -5.807545e-02  6.713523e+00  7.337707e+00
#>  [71]  4.701447e-01 -8.163291e-02 -4.636475e-01 -4.319503e-01 -2.487281e-02
#>  [76]  2.337703e-01 -1.559756e-01 -1.613665e-01  1.005546e-01  5.882674e-02
#>  [81] -1.617886e-01 -1.486448e-02  1.102940e+01  6.826856e+00  1.497196e+01
#>  [86] -5.097863e-02  9.697329e-02  7.913696e-04  4.999781e-02  4.764802e-04
#>  [91] -2.435474e-02 -4.378732e-01 -4.427953e-01 -1.353958e-01 -2.121243e-02
#>  [96]  7.028809e+01  1.798567e+00 -3.409253e-03 -2.894496e-01  2.062779e+00
#> [101]  8.780379e-03  9.196334e-02 -5.270888e-02  6.384543e-02  4.056640e-02
#> [106] -3.202612e-01 -2.125442e-01  9.716084e+00 -3.745068e-01  2.341499e+00
#> [111]  5.578865e-02  1.772334e-02 -4.075757e-01  1.469198e+01  3.614157e+01
#> [116] -1.967244e-01  6.535727e-01  9.875884e+01 -9.060053e-02 -3.205437e-01
#> [121] -4.085574e-01 -9.763332e-03 -2.095751e-01  1.599335e+02 -2.721192e-01
#> [126]  9.193069e-02  1.230001e-01  1.059611e-01 -1.897193e-01  5.725615e+00
#> [131] -4.157221e-01 -7.994792e-03  4.302137e-01  8.971552e-02 -3.705502e-01
#> [136] -4.822355e-01  5.236303e-02