Chi-squared test for multivariate outliers

Performs a Chi-squared test for multivariate outliers.

moutlier_chisq(
  xs,
  mask = !Reduce("|", lapply(xs, is.na)),
  threshold = c(0.9, 0.95),
  return.score = FALSE
)

Arguments

xs
A dataframe or list of vectors (which will be coerced to a numeric matrix).

mask
A logical vector that defines which values in x will used when computing statistics. Useful when a subset of quality-assured data is available. Default mask is non-NA Values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rnorm(length(x), mean = 1, sd = 3)
y = sin(x) + noise
mask = noise < 1

moutlier_chisq(list(x, y))

#>   [1] not outlier     extreme outlier not outlier     not outlier    
#>   [5] not outlier     not outlier     not outlier     not outlier    
#>   [9] not outlier     not outlier     not outlier     not outlier    
#>  [13] extreme outlier not outlier     not outlier     not outlier    
#>  [17] not outlier     not outlier     not outlier     not outlier    
#>  [21] not outlier     not outlier     not outlier     not outlier    
#>  [25] not outlier     not outlier     not outlier     extreme outlier
#>  [29] mild outlier    not outlier     not outlier     not outlier    
#>  [33] not outlier     not outlier     not outlier     not outlier    
#>  [37] not outlier     not outlier     not outlier     not outlier    
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     not outlier    
#>  [49] not outlier     not outlier     extreme outlier not outlier    
#>  [53] not outlier     not outlier     not outlier     not outlier    
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     not outlier     not outlier     not outlier    
#>  [65] not outlier     not outlier     extreme outlier not outlier    
#>  [69] not outlier     not outlier     not outlier     not outlier    
#>  [73] not outlier     not outlier     not outlier     not outlier    
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     not outlier     not outlier     not outlier    
#>  [85] not outlier     not outlier     not outlier     not outlier    
#>  [89] not outlier     not outlier     not outlier     mild outlier   
#>  [93] not outlier     not outlier     not outlier     not outlier    
#>  [97] not outlier     not outlier     not outlier     not outlier    
#> [101] not outlier     not outlier     not outlier     extreme outlier
#> [105] not outlier     not outlier     not outlier     not outlier    
#> [109] not outlier     not outlier     not outlier     not outlier    
#> [113] not outlier     not outlier     not outlier     extreme outlier
#> [117] not outlier     not outlier     not outlier     not outlier    
#> [121] not outlier     extreme outlier not outlier     not outlier    
#> [125] not outlier     not outlier     not outlier     not outlier    
#> [129] not outlier     mild outlier    not outlier     not outlier    
#> [133] not outlier     not outlier     not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

moutlier_chisq(list(x, y), mask)

#>   [1] not outlier     extreme outlier not outlier     not outlier    
#>   [5] not outlier     not outlier     not outlier     not outlier    
#>   [9] not outlier     not outlier     not outlier     not outlier    
#>  [13] extreme outlier not outlier     not outlier     not outlier    
#>  [17] not outlier     not outlier     not outlier     not outlier    
#>  [21] not outlier     not outlier     not outlier     not outlier    
#>  [25] not outlier     not outlier     not outlier     extreme outlier
#>  [29] mild outlier    not outlier     not outlier     not outlier    
#>  [33] not outlier     not outlier     not outlier     not outlier    
#>  [37] not outlier     not outlier     not outlier     not outlier    
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     not outlier    
#>  [49] not outlier     not outlier     extreme outlier not outlier    
#>  [53] not outlier     not outlier     not outlier     not outlier    
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     not outlier     not outlier     not outlier    
#>  [65] not outlier     not outlier     extreme outlier not outlier    
#>  [69] not outlier     not outlier     not outlier     not outlier    
#>  [73] not outlier     not outlier     not outlier     not outlier    
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     not outlier     not outlier     not outlier    
#>  [85] not outlier     not outlier     not outlier     not outlier    
#>  [89] not outlier     not outlier     not outlier     mild outlier   
#>  [93] not outlier     not outlier     not outlier     not outlier    
#>  [97] not outlier     not outlier     not outlier     not outlier    
#> [101] not outlier     not outlier     not outlier     extreme outlier
#> [105] not outlier     not outlier     not outlier     not outlier    
#> [109] not outlier     not outlier     not outlier     not outlier    
#> [113] not outlier     not outlier     not outlier     extreme outlier
#> [117] not outlier     not outlier     not outlier     not outlier    
#> [121] not outlier     extreme outlier not outlier     not outlier    
#> [125] not outlier     not outlier     not outlier     not outlier    
#> [129] not outlier     mild outlier    not outlier     not outlier    
#> [133] not outlier     not outlier     not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

moutlier_chisq(list(x, y), mask, threshold = c(0.8, 0.9))

#>   [1] not outlier     extreme outlier not outlier     not outlier    
#>   [5] not outlier     not outlier     not outlier     mild outlier   
#>   [9] not outlier     not outlier     not outlier     not outlier    
#>  [13] extreme outlier mild outlier    not outlier     not outlier    
#>  [17] not outlier     not outlier     not outlier     not outlier    
#>  [21] not outlier     not outlier     not outlier     not outlier    
#>  [25] not outlier     not outlier     not outlier     extreme outlier
#>  [29] extreme outlier not outlier     not outlier     not outlier    
#>  [33] not outlier     not outlier     not outlier     not outlier    
#>  [37] not outlier     not outlier     not outlier     not outlier    
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     not outlier    
#>  [49] not outlier     not outlier     extreme outlier not outlier    
#>  [53] not outlier     not outlier     not outlier     not outlier    
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     not outlier     not outlier     not outlier    
#>  [65] not outlier     not outlier     extreme outlier not outlier    
#>  [69] not outlier     not outlier     not outlier     not outlier    
#>  [73] not outlier     not outlier     not outlier     not outlier    
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     not outlier     not outlier     not outlier    
#>  [85] not outlier     not outlier     not outlier     not outlier    
#>  [89] not outlier     not outlier     not outlier     extreme outlier
#>  [93] not outlier     not outlier     not outlier     not outlier    
#>  [97] not outlier     mild outlier    not outlier     not outlier    
#> [101] not outlier     mild outlier    not outlier     extreme outlier
#> [105] not outlier     not outlier     not outlier     not outlier    
#> [109] not outlier     not outlier     mild outlier    not outlier    
#> [113] not outlier     not outlier     not outlier     extreme outlier
#> [117] not outlier     not outlier     not outlier     not outlier    
#> [121] not outlier     extreme outlier not outlier     not outlier    
#> [125] mild outlier    not outlier     not outlier     not outlier    
#> [129] not outlier     extreme outlier not outlier     not outlier    
#> [133] not outlier     mild outlier    not outlier     mild outlier   
#> [137] not outlier    
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

moutlier_chisq(list(x, y), return.score = TRUE)

#>   [1]  2.97617269  6.88395062  2.91250498  3.08848180  2.64592125  3.13065873
#>   [7]  2.59912931  3.43408078  2.81067788  2.32201602  2.30984209  2.38930098
#>  [13]  6.63303761  4.42588430  2.45658642  2.37405917  1.72677863  2.30484677
#>  [19]  2.50031621  1.52575245  2.00697060  1.75655674  1.76140668  3.11978990
#>  [25]  1.59607033  1.91678663  2.50492092  6.15250049  5.18152383  0.97645388
#>  [31]  1.21131595  1.26519986  2.71358925  0.82025856  0.75223064  1.33059301
#>  [37]  0.84551126  0.87458389  1.84218287  1.34270524  1.06030614  0.83666731
#>  [43]  0.80876843  1.24331300  0.36656579  0.62147449  0.51268903  0.99690869
#>  [49]  1.80398832  1.14249372 10.03426189  0.20696615  0.29299038  0.28120101
#>  [55]  1.53025195  0.68224241  0.11536922  1.02792825  2.21322885  0.23965542
#>  [61]  0.05532429  0.17110830  0.11052534  0.17246677  0.37242805  0.14600946
#>  [67]  6.06364159  0.38380672  0.52904425  0.71905120  0.14312226  0.47716710
#>  [73]  0.13404584  0.64894838  1.36548732  0.92578062  0.35187650  0.05502217
#>  [79]  0.85536403  2.94174816  0.09150272  3.06265278  0.17523091  0.15195488
#>  [85]  0.55252565  0.56911473  0.33262610  0.45441299  0.65852578  0.51498540
#>  [91]  0.61898969  4.98326063  0.55454779  0.40692512  1.61795222  0.46583704
#>  [97]  0.56754466  4.54094417  0.62914531  1.45811741  0.65576740  4.36577724
#> [103]  2.18114467  9.03924254  1.78262141  1.03301378  1.00523674  1.72483061
#> [109]  1.07579948  2.11391895  3.84635057  2.17438126  1.55486153  2.62410431
#> [115]  1.87789445  6.12424697  2.16238840  1.56782770  2.14138293  2.14291350
#> [121]  1.71658613  6.78130347  2.20235774  2.38392834  4.39230816  2.45149824
#> [127]  3.01050734  2.30014451  2.48464494  5.89641590  2.44726450  2.59088893
#> [133]  2.84605720  3.53279778  2.90311572  3.38905606  3.01023661