Isolation Forest outlier detection

Performs outlier detection using an Isolation Forest.

outlier_iforest(
  x,
  mask = !is.na(x),
  threshold = c(0.8, 0.9),
  return.score = FALSE,
  ...
)

Arguments

x
A vector of data.

mask
A logical mask that identifies a subgroup of x to compute non-parametric statistics on. Useful when a subset of quality-assured data is available. Default action is to ignore NA values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

...
Additional arguments to solitude::isolationForest$new(). note that the argument sample_size will be overwritten to use the number of unmasked data points, i.e. length(which(mask)).

Details

the values of threshold identify mild and extreme\ outliers based on the Isolation Forest score in the range [0,1]. Default values are 0.8 for "mild" outliers and 0.9 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

outlier_iforest(y)

#> Loading required namespace: solitude

#> INFO  [21:19:09.573] Building Isolation Forest ...  
#> INFO  [21:19:10.828] done 
#> INFO  [21:19:10.839] Computing depth of terminal nodes ...  
#> INFO  [21:19:11.655] done 
#> INFO  [21:19:11.668] Completed growing isolation forest 

#>   [1] not outlier     not outlier     not outlier     not outlier    
#>   [5] not outlier     not outlier     not outlier     not outlier    
#>   [9] not outlier     not outlier     not outlier     not outlier    
#>  [13] not outlier     not outlier     not outlier     not outlier    
#>  [17] not outlier     not outlier     not outlier     extreme outlier
#>  [21] not outlier     not outlier     not outlier     not outlier    
#>  [25] not outlier     not outlier     not outlier     not outlier    
#>  [29] not outlier     not outlier     not outlier     not outlier    
#>  [33] not outlier     not outlier     not outlier     not outlier    
#>  [37] not outlier     not outlier     not outlier     not outlier    
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     not outlier    
#>  [49] not outlier     not outlier     not outlier     not outlier    
#>  [53] not outlier     not outlier     not outlier     not outlier    
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     not outlier     not outlier     not outlier    
#>  [65] not outlier     not outlier     not outlier     not outlier    
#>  [69] not outlier     not outlier     not outlier     not outlier    
#>  [73] not outlier     not outlier     not outlier     not outlier    
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     not outlier     not outlier     not outlier    
#>  [85] not outlier     not outlier     not outlier     not outlier    
#>  [89] not outlier     not outlier     not outlier     not outlier    
#>  [93] not outlier     not outlier     not outlier     not outlier    
#>  [97] not outlier     not outlier     not outlier     not outlier    
#> [101] not outlier     not outlier     not outlier     not outlier    
#> [105] not outlier     not outlier     not outlier     not outlier    
#> [109] not outlier     not outlier     not outlier     mild outlier   
#> [113] not outlier     not outlier     not outlier     not outlier    
#> [117] not outlier     not outlier     not outlier     not outlier    
#> [121] not outlier     not outlier     not outlier     not outlier    
#> [125] not outlier     not outlier     not outlier     not outlier    
#> [129] not outlier     not outlier     not outlier     not outlier    
#> [133] not outlier     not outlier     not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < mild outlier < extreme outlier

outlier_iforest(y, mask)

#> INFO  [21:19:11.695] Building Isolation Forest ...  
#> INFO  [21:19:11.701] done 
#> INFO  [21:19:11.703] Computing depth of terminal nodes ...  
#> INFO  [21:19:12.005] done 
#> INFO  [21:19:12.013] Completed growing isolation forest 

#>   [1] not outlier not outlier not outlier not outlier not outlier not outlier
#>   [7] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [13] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [19] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [25] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [31] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [37] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [43] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [49] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [55] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [61] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [67] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [73] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [79] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [85] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [91] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [97] not outlier not outlier not outlier not outlier not outlier not outlier
#> [103] not outlier not outlier not outlier not outlier not outlier not outlier
#> [109] not outlier not outlier not outlier not outlier not outlier not outlier
#> [115] not outlier not outlier not outlier not outlier not outlier not outlier
#> [121] not outlier not outlier not outlier not outlier not outlier not outlier
#> [127] not outlier not outlier not outlier not outlier not outlier not outlier
#> [133] not outlier not outlier not outlier not outlier not outlier
#> Levels: not outlier < mild outlier < extreme outlier

outlier_iforest(y, mask, threshold = c(1, 2))

#> INFO  [21:19:12.034] Building Isolation Forest ...  
#> INFO  [21:19:12.041] done 
#> INFO  [21:19:12.043] Computing depth of terminal nodes ...  
#> INFO  [21:19:12.313] done 
#> INFO  [21:19:12.321] Completed growing isolation forest 

#>   [1] not outlier not outlier not outlier not outlier not outlier not outlier
#>   [7] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [13] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [19] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [25] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [31] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [37] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [43] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [49] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [55] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [61] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [67] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [73] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [79] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [85] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [91] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [97] not outlier not outlier not outlier not outlier not outlier not outlier
#> [103] not outlier not outlier not outlier not outlier not outlier not outlier
#> [109] not outlier not outlier not outlier not outlier not outlier not outlier
#> [115] not outlier not outlier not outlier not outlier not outlier not outlier
#> [121] not outlier not outlier not outlier not outlier not outlier not outlier
#> [127] not outlier not outlier not outlier not outlier not outlier not outlier
#> [133] not outlier not outlier not outlier not outlier not outlier
#> Levels: not outlier < mild outlier < extreme outlier

outlier_iforest(y, return.score = TRUE)

#> INFO  [21:19:12.341] Building Isolation Forest ...  
#> INFO  [21:19:12.347] done 
#> INFO  [21:19:12.348] Computing depth of terminal nodes ...  
#> INFO  [21:19:12.615] done 
#> INFO  [21:19:12.625] Completed growing isolation forest 

#>   [1] 0.5398205 0.5398205 0.5406532 0.5410700 0.5402367 0.5410700 0.5398205
#>   [8] 0.5799298 0.5402367 0.5398205 0.5406532 0.5406532 0.5402367 0.5398205
#>  [15] 0.5398205 0.5410700 0.5662394 0.5593003 0.5398205 0.9151891 0.5398205
#>  [22] 0.5410700 0.5402367 0.5398205 0.5688637 0.5398205 0.5398205 0.5653673
#>  [29] 0.5398205 0.5398205 0.5435777 0.5419046 0.5419046 0.5402367 0.5398205
#>  [36] 0.5398205 0.5406532 0.5410700 0.5398205 0.5398205 0.5398205 0.5435777
#>  [43] 0.5410700 0.5465179 0.5398205 0.5398205 0.5406532 0.5593003 0.6590732
#>  [50] 0.5419046 0.5866725 0.5402367 0.5398205 0.5406532 0.5398205 0.5410700
#>  [57] 0.5419046 0.5398205 0.5398205 0.5406532 0.6750081 0.5410700 0.5402367
#>  [64] 0.5402367 0.5444161 0.5398205 0.6977512 0.5398205 0.5419046 0.5398205
#>  [71] 0.5532984 0.5994694 0.5398205 0.5398205 0.5398205 0.5398205 0.5406532
#>  [78] 0.5410700 0.5439967 0.5402367 0.5398205 0.5402367 0.5511705 0.5406532
#>  [85] 0.6435145 0.5398205 0.5402367 0.5398205 0.5398205 0.5402367 0.5406532
#>  [92] 0.5410700 0.5419046 0.6003941 0.5444161 0.5588695 0.5398205 0.5402367
#>  [99] 0.6667361 0.5398205 0.5410700 0.5402367 0.5398205 0.5398205 0.5398205
#> [106] 0.5406532 0.5460969 0.5398205 0.5402367 0.5406532 0.5439967 0.8121461
#> [113] 0.5398205 0.5398205 0.5398205 0.5410700 0.5398205 0.5410700 0.5456762
#> [120] 0.5398205 0.5406532 0.5402367 0.5398205 0.5406532 0.6069068 0.5402367
#> [127] 0.5423224 0.5398205 0.5697411 0.5398205 0.5614596 0.5452559 0.5398205
#> [134] 0.5402367 0.5398205 0.5486279 0.5406532