LOF outlier detection

Performs outlier detection using Local Outlier Factor algorithm.

outlier_lof(
  x,
  mask = !is.na(x),
  threshold = c(1.5, 2),
  return.score = FALSE,
  ...
)

Arguments

x
A vector of data.

mask
A logical mask that identifies a subgroup of x to compute non-parametric statistics on. Useful when a subset of quality-assured data is available. Default action is to ignore NA values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

...
Additional arguments to dbscan::lof, namely k.

Details

the values of threshold identify mild and extreme outliers based on the LOF score. Scores significantly larger than 1 indicate outliers. Default values are 1.5 for "mild" outliers and 2.0 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

outlier_lof(y)

#> Loading required namespace: dbscan

#>   [1] not outlier     not outlier     not outlier     not outlier    
#>   [5] mild outlier    mild outlier    not outlier     not outlier    
#>   [9] not outlier     not outlier     not outlier     not outlier    
#>  [13] not outlier     not outlier     mild outlier    not outlier    
#>  [17] not outlier     not outlier     not outlier     not outlier    
#>  [21] not outlier     not outlier     mild outlier    not outlier    
#>  [25] not outlier     not outlier     not outlier     not outlier    
#>  [29] mild outlier    not outlier     not outlier     not outlier    
#>  [33] not outlier     not outlier     not outlier     not outlier    
#>  [37] extreme outlier not outlier     not outlier     not outlier    
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     not outlier    
#>  [49] not outlier     not outlier     not outlier     not outlier    
#>  [53] mild outlier    extreme outlier not outlier     not outlier    
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     not outlier     not outlier     mild outlier   
#>  [65] not outlier     not outlier     not outlier     not outlier    
#>  [69] not outlier     not outlier     mild outlier    extreme outlier
#>  [73] not outlier     not outlier     not outlier     extreme outlier
#>  [77] not outlier     not outlier     mild outlier    not outlier    
#>  [81] not outlier     not outlier     not outlier     not outlier    
#>  [85] not outlier     not outlier     not outlier     not outlier    
#>  [89] not outlier     not outlier     not outlier     not outlier    
#>  [93] extreme outlier not outlier     not outlier     extreme outlier
#>  [97] not outlier     mild outlier    not outlier     mild outlier   
#> [101] not outlier     not outlier     not outlier     not outlier    
#> [105] not outlier     not outlier     mild outlier    not outlier    
#> [109] not outlier     mild outlier    mild outlier    not outlier    
#> [113] not outlier     not outlier     not outlier     extreme outlier
#> [117] not outlier     not outlier     mild outlier    extreme outlier
#> [121] not outlier     not outlier     not outlier     mild outlier   
#> [125] not outlier     not outlier     not outlier     not outlier    
#> [129] mild outlier    not outlier     mild outlier    not outlier    
#> [133] not outlier     not outlier     not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < mild outlier < extreme outlier

outlier_lof(y, mask)

#>   [1] not outlier     not outlier     mild outlier    mild outlier   
#>   [5] not outlier     not outlier     not outlier     not outlier    
#>   [9] not outlier     not outlier     not outlier     not outlier    
#>  [13] not outlier     not outlier     not outlier     not outlier    
#>  [17] not outlier     mild outlier    not outlier     not outlier    
#>  [21] not outlier     not outlier     not outlier     not outlier    
#>  [25] not outlier     not outlier     not outlier     not outlier    
#>  [29] not outlier     not outlier     not outlier     not outlier    
#>  [33] mild outlier    not outlier     not outlier     not outlier    
#>  [37] not outlier     not outlier     not outlier     not outlier    
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     not outlier    
#>  [49] not outlier     mild outlier    extreme outlier not outlier    
#>  [53] not outlier     not outlier     not outlier     not outlier    
#>  [57] not outlier     not outlier     mild outlier    not outlier    
#>  [61] not outlier     not outlier     not outlier     not outlier    
#>  [65] not outlier     not outlier     not outlier     not outlier    
#>  [69] not outlier     not outlier     not outlier     not outlier    
#>  [73] not outlier     not outlier     mild outlier    extreme outlier
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     not outlier     not outlier     not outlier    
#>  [85] not outlier     not outlier     not outlier     not outlier    
#>  [89] not outlier     mild outlier    not outlier     not outlier    
#>  [93] not outlier     not outlier     not outlier     not outlier    
#>  [97] not outlier     mild outlier    not outlier     mild outlier   
#> [101] not outlier     not outlier     not outlier     not outlier    
#> [105] not outlier     not outlier     mild outlier    not outlier    
#> [109] not outlier     not outlier     not outlier     not outlier    
#> [113] not outlier     not outlier     not outlier     not outlier    
#> [117] not outlier     not outlier     not outlier     not outlier    
#> [121] not outlier     not outlier     not outlier     not outlier    
#> [125] not outlier     not outlier     not outlier     not outlier    
#> [129] not outlier     not outlier     not outlier     not outlier    
#> [133] not outlier     not outlier     not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < mild outlier < extreme outlier

outlier_lof(y, mask, threshold = c(1, 2))

#>   [1] mild outlier    mild outlier    mild outlier    mild outlier   
#>   [5] mild outlier    not outlier     not outlier     not outlier    
#>   [9] not outlier     not outlier     mild outlier    mild outlier   
#>  [13] not outlier     not outlier     not outlier     not outlier    
#>  [17] not outlier     mild outlier    not outlier     not outlier    
#>  [21] not outlier     mild outlier    not outlier     mild outlier   
#>  [25] not outlier     not outlier     not outlier     not outlier    
#>  [29] not outlier     mild outlier    not outlier     mild outlier   
#>  [33] mild outlier    not outlier     not outlier     not outlier    
#>  [37] not outlier     not outlier     not outlier     mild outlier   
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     mild outlier   
#>  [49] not outlier     mild outlier    extreme outlier not outlier    
#>  [53] mild outlier    not outlier     not outlier     not outlier    
#>  [57] not outlier     mild outlier    mild outlier    mild outlier   
#>  [61] not outlier     not outlier     mild outlier    not outlier    
#>  [65] not outlier     not outlier     not outlier     not outlier    
#>  [69] mild outlier    not outlier     not outlier     not outlier    
#>  [73] not outlier     not outlier     mild outlier    extreme outlier
#>  [77] not outlier     not outlier     not outlier     mild outlier   
#>  [81] not outlier     mild outlier    not outlier     not outlier    
#>  [85] mild outlier    not outlier     not outlier     not outlier    
#>  [89] mild outlier    mild outlier    not outlier     not outlier    
#>  [93] not outlier     not outlier     not outlier     not outlier    
#>  [97] not outlier     mild outlier    not outlier     mild outlier   
#> [101] not outlier     not outlier     not outlier     mild outlier   
#> [105] not outlier     not outlier     mild outlier    mild outlier   
#> [109] not outlier     not outlier     not outlier     not outlier    
#> [113] mild outlier    not outlier     mild outlier    not outlier    
#> [117] not outlier     not outlier     not outlier     not outlier    
#> [121] not outlier     not outlier     not outlier     not outlier    
#> [125] not outlier     not outlier     not outlier     not outlier    
#> [129] not outlier     not outlier     not outlier     not outlier    
#> [133] not outlier     mild outlier    not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < mild outlier < extreme outlier

outlier_lof(y, return.score = TRUE)

#>   [1]  1.0828501  0.9849167  0.9613808  1.2014796  1.6164007  1.5807164
#>   [7]  1.0058533  1.0133167  1.1563932  1.0110473  0.9448089  0.9367136
#>  [13]  1.1073547  0.9070709  1.6969236  1.1819839  0.9826209  1.2014796
#>  [19]  0.9826209  1.0582541  1.0248121  1.0769525  1.9881511  1.2311038
#>  [25]  1.1071850  1.0419946  0.9078279  0.9571434  1.7353696  1.1684665
#>  [31]  1.2110418  1.0120753  1.4809488  1.1267769  0.9432792  1.0024078
#>  [37]  4.4937141  0.9935460  0.9452044  1.2305754  1.1114487  1.0900893
#>  [43]  0.9922693  1.0813390  0.9950682  1.0501091  0.9768079  1.1047280
#>  [49]  1.0240229  1.2014796  1.3736187  0.9809034  1.8684626  4.3009242
#>  [55]  1.1114487  0.9613808  1.4494470  1.0150138  1.0238392  1.0386941
#>  [61]  1.1563932  1.1734381  1.4530157  1.8365342  0.9674765  0.9416990
#>  [67]  0.9808865  0.9318016  1.0149813  1.1701984  1.5116676  2.4214728
#>  [73]  1.0864542  0.9416990  1.0446729  2.0687044  1.0283342  0.9569077
#>  [79]  1.6399976  1.0390922  1.4494470  1.0828832  0.9685280  0.9912541
#>  [85]  1.0398758  0.9768932  1.1819839  0.9569077  0.9783607  1.2014796
#>  [91]  1.1830693  0.9064036  4.4288001  1.0419946  0.9366570 33.9648547
#>  [97]  1.0738806  1.8457164  0.9366570  1.8475692  0.9318016  0.8885728
#> [103]  1.2627863  1.1422980  1.1164730  0.9766162  1.5349466  0.9673113
#> [109]  1.0497623  1.5727300  1.7396920  1.0211210  1.0104935  1.0075439
#> [115]  0.9831949  6.6581652  0.8993676  0.9538680  1.7396920  2.1560195
#> [121]  1.0442964  0.9517712  1.0633868  1.8091547  1.0073891  1.4512876
#> [127]  1.0158768  1.3802857  1.7396920  1.0116874  1.7396920  1.0296497
#> [133]  0.9867166  1.1971078  0.9309406  1.0240229  0.9851134