Median absolute deviation test for outliers

Performs a median absolute deviation (MAD) test for outliers.

outlier_mad(x, mask = !is.na(x), threshold = c(1.5, 3), return.score = FALSE)

Arguments

x
A vector of data.

mask
A logical mask that identifies a subgroup of x to compute non-parametric statistics on. Useful when a subset of quality-assured data is available. Default action is to ignore NA values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

Details

the values of threshold identify the multiplier of the median distance used to identify mild and extreme outliers. Default values are 1.5 for "mild" outliers and 3.0 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

outlier_mad(y)

#>   [1] not outlier     not outlier     not outlier     not outlier    
#>   [5] not outlier     mild outlier    not outlier     not outlier    
#>   [9] not outlier     not outlier     not outlier     not outlier    
#>  [13] not outlier     not outlier     mild outlier    not outlier    
#>  [17] not outlier     not outlier     extreme outlier not outlier    
#>  [21] not outlier     not outlier     mild outlier    not outlier    
#>  [25] not outlier     not outlier     not outlier     not outlier    
#>  [29] mild outlier    mild outlier    not outlier     not outlier    
#>  [33] not outlier     not outlier     not outlier     not outlier    
#>  [37] not outlier     not outlier     mild outlier    not outlier    
#>  [41] not outlier     not outlier     not outlier     extreme outlier
#>  [45] extreme outlier mild outlier    extreme outlier mild outlier   
#>  [49] not outlier     not outlier     extreme outlier mild outlier   
#>  [53] extreme outlier mild outlier    mild outlier    not outlier    
#>  [57] not outlier     extreme outlier not outlier     not outlier    
#>  [61] not outlier     not outlier     mild outlier    extreme outlier
#>  [65] not outlier     mild outlier    not outlier     not outlier    
#>  [69] not outlier     extreme outlier mild outlier    not outlier    
#>  [73] not outlier     not outlier     not outlier     not outlier    
#>  [77] not outlier     mild outlier    mild outlier    not outlier    
#>  [81] not outlier     not outlier     not outlier     extreme outlier
#>  [85] not outlier     not outlier     mild outlier    mild outlier   
#>  [89] not outlier     not outlier     not outlier     mild outlier   
#>  [93] extreme outlier mild outlier    mild outlier    mild outlier   
#>  [97] mild outlier    not outlier     not outlier     not outlier    
#> [101] extreme outlier not outlier     extreme outlier not outlier    
#> [105] extreme outlier not outlier     not outlier     mild outlier   
#> [109] not outlier     not outlier     mild outlier    not outlier    
#> [113] not outlier     extreme outlier extreme outlier not outlier    
#> [117] extreme outlier not outlier     extreme outlier mild outlier   
#> [121] not outlier     not outlier     not outlier     not outlier    
#> [125] not outlier     extreme outlier extreme outlier extreme outlier
#> [129] not outlier     not outlier     not outlier     not outlier    
#> [133] not outlier     extreme outlier not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < mild outlier < extreme outlier

outlier_mad(y, mask)

#>   [1] not outlier     mild outlier    mild outlier    extreme outlier
#>   [5] extreme outlier not outlier     not outlier     extreme outlier
#>   [9] extreme outlier extreme outlier not outlier     not outlier    
#>  [13] not outlier     extreme outlier mild outlier    extreme outlier
#>  [17] not outlier     extreme outlier extreme outlier not outlier    
#>  [21] mild outlier    extreme outlier not outlier     extreme outlier
#>  [25] extreme outlier not outlier     extreme outlier not outlier    
#>  [29] extreme outlier not outlier     extreme outlier extreme outlier
#>  [33] not outlier     extreme outlier mild outlier    extreme outlier
#>  [37] extreme outlier mild outlier    mild outlier    not outlier    
#>  [41] not outlier     extreme outlier mild outlier    extreme outlier
#>  [45] extreme outlier not outlier     extreme outlier not outlier    
#>  [49] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [53] extreme outlier not outlier     not outlier     extreme outlier
#>  [57] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [61] not outlier     not outlier     mild outlier    extreme outlier
#>  [65] extreme outlier extreme outlier mild outlier    mild outlier   
#>  [69] not outlier     extreme outlier not outlier     mild outlier   
#>  [73] mild outlier    not outlier     mild outlier    not outlier    
#>  [77] mild outlier    not outlier     not outlier     extreme outlier
#>  [81] extreme outlier mild outlier    extreme outlier extreme outlier
#>  [85] extreme outlier extreme outlier not outlier     not outlier    
#>  [89] not outlier     not outlier     mild outlier    extreme outlier
#>  [93] extreme outlier not outlier     mild outlier    extreme outlier
#>  [97] extreme outlier not outlier     not outlier     not outlier    
#> [101] extreme outlier extreme outlier extreme outlier not outlier    
#> [105] extreme outlier extreme outlier extreme outlier extreme outlier
#> [109] not outlier     extreme outlier not outlier     not outlier    
#> [113] extreme outlier extreme outlier extreme outlier mild outlier   
#> [117] extreme outlier mild outlier    extreme outlier not outlier    
#> [121] not outlier     mild outlier    not outlier     mild outlier   
#> [125] extreme outlier extreme outlier extreme outlier extreme outlier
#> [129] not outlier     not outlier     mild outlier    not outlier    
#> [133] not outlier     extreme outlier not outlier     not outlier    
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

outlier_mad(y, mask, threshold = c(1, 2))

#>   [1] mild outlier    extreme outlier mild outlier    extreme outlier
#>   [5] extreme outlier mild outlier    not outlier     extreme outlier
#>   [9] extreme outlier extreme outlier mild outlier    not outlier    
#>  [13] not outlier     extreme outlier mild outlier    extreme outlier
#>  [17] not outlier     extreme outlier extreme outlier mild outlier   
#>  [21] mild outlier    extreme outlier not outlier     extreme outlier
#>  [25] extreme outlier not outlier     extreme outlier mild outlier   
#>  [29] extreme outlier not outlier     extreme outlier extreme outlier
#>  [33] not outlier     extreme outlier extreme outlier extreme outlier
#>  [37] extreme outlier mild outlier    mild outlier    mild outlier   
#>  [41] not outlier     extreme outlier extreme outlier extreme outlier
#>  [45] extreme outlier not outlier     extreme outlier mild outlier   
#>  [49] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [53] extreme outlier not outlier     mild outlier    extreme outlier
#>  [57] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [61] mild outlier    mild outlier    mild outlier    extreme outlier
#>  [65] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [69] not outlier     extreme outlier mild outlier    extreme outlier
#>  [73] mild outlier    not outlier     mild outlier    not outlier    
#>  [77] mild outlier    mild outlier    mild outlier    extreme outlier
#>  [81] extreme outlier mild outlier    extreme outlier extreme outlier
#>  [85] extreme outlier extreme outlier mild outlier    not outlier    
#>  [89] not outlier     not outlier     mild outlier    extreme outlier
#>  [93] extreme outlier mild outlier    mild outlier    extreme outlier
#>  [97] extreme outlier mild outlier    mild outlier    not outlier    
#> [101] extreme outlier extreme outlier extreme outlier mild outlier   
#> [105] extreme outlier extreme outlier extreme outlier extreme outlier
#> [109] not outlier     extreme outlier not outlier     not outlier    
#> [113] extreme outlier extreme outlier extreme outlier mild outlier   
#> [117] extreme outlier mild outlier    extreme outlier mild outlier   
#> [121] not outlier     extreme outlier mild outlier    mild outlier   
#> [125] extreme outlier extreme outlier extreme outlier extreme outlier
#> [129] not outlier     not outlier     mild outlier    not outlier    
#> [133] not outlier     extreme outlier not outlier     not outlier    
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

outlier_mad(y, return.score = TRUE)

#>   [1]  -0.688542849  -0.066469250  -0.398809337   0.189681811   0.383283612
#>   [6]  -1.732516950  -1.472385677   0.163013136   0.099932572   0.182502464
#>  [11]  -0.686500264  -1.000000000  -1.402780754   0.814011087  -1.967397275
#>  [16]   0.538702753  -1.365771394   0.626703961  12.160926809  -0.826305912
#>  [21]  -0.397020205   0.223596617  -1.680157635   0.829288094   0.126900936
#>  [26]  -0.942925589   0.738500474  -0.738064563   2.121956909  -1.722946551
#>  [31]   0.456508355   1.484700644  -1.410041148   0.602707677           NaN
#>  [36]   0.424523489   1.412044716  -0.396171420  -1.952912132  -0.712702081
#>  [41]  -1.392495029   0.148268137   0.004163127   5.262844517   6.997663546
#>  [46]  -1.554070223  10.585932429  -1.755155261   1.000000000   0.320355377
#>  [51]   9.163696242   2.991349487 974.303075357  -1.647783122  -1.809236956
#>  [56]   0.163871581   0.890279471   3.001384566   0.143080380  -0.535380961
#>  [61]  -0.814807638  -0.651055953  -1.964626634  69.240733870   0.052522461
#>  [66]   2.464136170   0.016722122  -0.033782471  -1.352275487   7.857014354
#>  [71]  -1.859631856   0.010308990  -0.561465925  -0.852510728  -0.394868300
#>  [76]  -0.918123094  -0.386592140  -1.794123465  -1.756447076   0.067260292
#>  [81]   0.599491769  -0.548587384   1.429915377   4.477574218   0.101294036
#>  [86]   0.260433666  -1.758243765  -1.684005289  -1.121921737  -1.002918174
#>  [91]  -0.497514791   1.896452106 909.441903093  -1.804160488  -1.975048495
#>  [96]   2.132681895   1.571342354  -0.729165172  -0.746447665  -0.953948627
#> [101]   8.151479868   1.160977341   7.537389016  -0.818069761 853.079325285
#> [106]   0.793697051   0.094136558   1.941338305  -1.218620695   0.391075355
#> [111]  -1.693225821  -1.037762665   0.463674406  25.268674828   9.878460213
#> [116]  -0.436644856   3.289964302  -0.498438230 110.736826251  -1.798998879
#> [121]  -1.396044260  -0.317571116  -0.833134247  -0.500344643   1.341207344
#> [126]  75.400013690  55.636749448  20.850955014  -1.270093645  -0.992665399
#> [131]  -0.423147531  -0.971361015  -1.321706359   5.116766444  -1.385638577
#> [136]  -1.372486788   0.257107206