t-score test for outliers

Performs a t-score test for outliers.

outlier_tscore(
  x,
  mask = !is.na(x),
  threshold = c(0.9, 0.95),
  return.score = FALSE
)

Arguments

x
A vector of data.

mask
A logical vector that defines which values in x will used when computing statistics. Useful when a subset of quality-assured data is available. Default mask is non-NA Values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

Details

The values of threshold identify the quantiles of the t-distribution used to identify mild and extreme outliers. Default values are 0.9 for "mild" outliers and 0.95 for "extreme" outliers.

The t-score is equivalent to the z-score for sample sizes greater than 30.

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rnorm(length(x), mean = 1, sd = 3)
y = sin(x) + noise
mask = noise < 1

outlier_tscore(y)

#>   [1] not outlier     not outlier     not outlier     not outlier    
#>   [5] not outlier     extreme outlier not outlier     not outlier    
#>   [9] extreme outlier extreme outlier not outlier     not outlier    
#>  [13] extreme outlier not outlier     not outlier     not outlier    
#>  [17] extreme outlier not outlier     not outlier     mild outlier   
#>  [21] not outlier     mild outlier    extreme outlier extreme outlier
#>  [25] not outlier     not outlier     mild outlier    not outlier    
#>  [29] not outlier     not outlier     not outlier     not outlier    
#>  [33] mild outlier    not outlier     not outlier     not outlier    
#>  [37] not outlier     not outlier     not outlier     not outlier    
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     not outlier    
#>  [49] not outlier     not outlier     not outlier     not outlier    
#>  [53] not outlier     not outlier     mild outlier    mild outlier   
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     not outlier     not outlier     not outlier    
#>  [65] not outlier     not outlier     not outlier     not outlier    
#>  [69] not outlier     not outlier     extreme outlier not outlier    
#>  [73] not outlier     not outlier     not outlier     not outlier    
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     not outlier     not outlier     not outlier    
#>  [85] mild outlier    not outlier     not outlier     not outlier    
#>  [89] not outlier     mild outlier    not outlier     extreme outlier
#>  [93] not outlier     not outlier     not outlier     extreme outlier
#>  [97] not outlier     not outlier     extreme outlier not outlier    
#> [101] not outlier     not outlier     not outlier     not outlier    
#> [105] not outlier     not outlier     extreme outlier not outlier    
#> [109] not outlier     not outlier     not outlier     not outlier    
#> [113] not outlier     not outlier     mild outlier    not outlier    
#> [117] mild outlier    not outlier     not outlier     not outlier    
#> [121] not outlier     not outlier     not outlier     not outlier    
#> [125] not outlier     not outlier     not outlier     not outlier    
#> [129] extreme outlier extreme outlier not outlier     not outlier    
#> [133] not outlier     not outlier     not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

outlier_tscore(y, mask)

#>   [1] not outlier     not outlier     mild outlier    not outlier    
#>   [5] not outlier     extreme outlier not outlier     extreme outlier
#>   [9] extreme outlier extreme outlier mild outlier    extreme outlier
#>  [13] mild outlier    not outlier     not outlier     not outlier    
#>  [17] extreme outlier mild outlier    extreme outlier extreme outlier
#>  [21] extreme outlier mild outlier    mild outlier    extreme outlier
#>  [25] mild outlier    extreme outlier extreme outlier not outlier    
#>  [29] not outlier     not outlier     not outlier     not outlier    
#>  [33] not outlier     not outlier     not outlier     not outlier    
#>  [37] extreme outlier extreme outlier not outlier     extreme outlier
#>  [41] extreme outlier not outlier     extreme outlier not outlier    
#>  [45] extreme outlier not outlier     not outlier     not outlier    
#>  [49] extreme outlier extreme outlier extreme outlier not outlier    
#>  [53] mild outlier    not outlier     not outlier     extreme outlier
#>  [57] extreme outlier extreme outlier extreme outlier not outlier    
#>  [61] extreme outlier not outlier     not outlier     extreme outlier
#>  [65] not outlier     mild outlier    not outlier     not outlier    
#>  [69] not outlier     not outlier     extreme outlier not outlier    
#>  [73] not outlier     extreme outlier extreme outlier mild outlier   
#>  [77] mild outlier    not outlier     not outlier     extreme outlier
#>  [81] extreme outlier not outlier     not outlier     not outlier    
#>  [85] not outlier     extreme outlier not outlier     extreme outlier
#>  [89] extreme outlier not outlier     not outlier     extreme outlier
#>  [93] mild outlier    not outlier     not outlier     extreme outlier
#>  [97] not outlier     mild outlier    extreme outlier extreme outlier
#> [101] not outlier     extreme outlier not outlier     not outlier    
#> [105] mild outlier    mild outlier    extreme outlier extreme outlier
#> [109] not outlier     not outlier     not outlier     not outlier    
#> [113] not outlier     extreme outlier extreme outlier extreme outlier
#> [117] extreme outlier not outlier     extreme outlier not outlier    
#> [121] not outlier     not outlier     extreme outlier not outlier    
#> [125] extreme outlier not outlier     mild outlier    not outlier    
#> [129] extreme outlier extreme outlier not outlier     mild outlier   
#> [133] not outlier     not outlier     not outlier     not outlier    
#> [137] not outlier    
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

outlier_tscore(y, mask, threshold = c(0.8, 0.9))

#>   [1] not outlier     mild outlier    extreme outlier not outlier    
#>   [5] not outlier     extreme outlier not outlier     extreme outlier
#>   [9] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [13] extreme outlier mild outlier    not outlier     not outlier    
#>  [17] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [21] extreme outlier extreme outlier extreme outlier extreme outlier
#>  [25] extreme outlier extreme outlier extreme outlier not outlier    
#>  [29] mild outlier    not outlier     not outlier     not outlier    
#>  [33] mild outlier    not outlier     not outlier     not outlier    
#>  [37] extreme outlier extreme outlier mild outlier    extreme outlier
#>  [41] extreme outlier mild outlier    extreme outlier not outlier    
#>  [45] extreme outlier not outlier     mild outlier    not outlier    
#>  [49] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [53] extreme outlier not outlier     mild outlier    extreme outlier
#>  [57] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [61] extreme outlier mild outlier    mild outlier    extreme outlier
#>  [65] not outlier     extreme outlier mild outlier    not outlier    
#>  [69] not outlier     not outlier     extreme outlier not outlier    
#>  [73] not outlier     extreme outlier extreme outlier extreme outlier
#>  [77] extreme outlier mild outlier    not outlier     extreme outlier
#>  [81] extreme outlier mild outlier    mild outlier    not outlier    
#>  [85] mild outlier    extreme outlier not outlier     extreme outlier
#>  [89] extreme outlier mild outlier    not outlier     extreme outlier
#>  [93] extreme outlier mild outlier    not outlier     extreme outlier
#>  [97] mild outlier    extreme outlier extreme outlier extreme outlier
#> [101] not outlier     extreme outlier mild outlier    not outlier    
#> [105] extreme outlier extreme outlier extreme outlier extreme outlier
#> [109] mild outlier    not outlier     not outlier     mild outlier   
#> [113] mild outlier    extreme outlier extreme outlier extreme outlier
#> [117] extreme outlier not outlier     extreme outlier mild outlier   
#> [121] not outlier     not outlier     extreme outlier not outlier    
#> [125] extreme outlier mild outlier    extreme outlier mild outlier   
#> [129] extreme outlier extreme outlier not outlier     extreme outlier
#> [133] not outlier     not outlier     not outlier     mild outlier   
#> [137] not outlier    
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

outlier_tscore(y, return.score = TRUE)

#>   [1] -1.140952676  0.022691422  0.143951656 -0.265653619 -0.445874704
#>   [6] -2.138440672 -0.769285373  0.532531909 -1.909879346  1.953943936
#>  [11]  0.201936664  0.766271696 -1.674214471  0.031470455 -0.676309621
#>  [16] -0.976519801 -2.378742962  0.130324197  0.516007070  1.397296769
#>  [21]  0.908375829 -1.628993395 -1.690461634 -2.269005279  0.197763869
#>  [26]  1.076663734  1.558001616 -1.094797353 -0.132002155 -0.334578176
#>  [31] -0.812464124 -1.084296116 -1.337210108 -0.728953111 -0.472397315
#>  [36] -1.165001185  0.669531641  0.498739358 -0.109867902  1.016220526
#>  [41]  1.271514813 -0.194879262  0.686706489 -0.530413149  0.952197554
#>  [46] -0.226546585  0.018043715 -0.476304734  0.948092229  0.396827774
#>  [51]  1.167785259 -0.003663835  0.215285183 -0.911874680 -1.579098418
#>  [56]  1.440528952  0.404290354  0.865829777  0.470175187 -0.091380259
#>  [61]  0.350865776  0.055543737 -0.009277496  1.286268603 -0.765996766
#>  [66]  0.129104098  0.015163991 -0.881862246 -0.725256600 -0.908981564
#>  [71] -2.093697197 -0.795487529 -1.218635765  0.607784186  1.085847599
#>  [76]  0.081721568  0.245174178  0.052904899 -0.226866612  0.532868504
#>  [81]  0.708374411  0.003869021 -0.032036631 -0.274875985 -1.427923191
#>  [86]  0.869891754 -0.318502770  1.157265449  1.018904244 -1.515246651
#>  [91] -0.538448499  2.439872780  0.235509875 -0.107867874 -0.854584441
#>  [96]  2.301082954 -0.182424671  0.150052396  1.668292438  0.923613877
#> [101] -1.225506475  1.170712580  0.008371573 -1.199896403  0.197371313
#> [106]  0.271918714  2.454374286  1.151023171 -0.092578867 -0.977203407
#> [111] -0.814232432  0.022235523 -0.029734444  0.400800854  1.446995965
#> [116]  0.417269945  1.378183310 -0.423044566  0.401309702  0.016597844
#> [121] -0.894774260 -0.975336013  0.683233173 -0.271688704  0.736092592
#> [126]  0.034536196  0.194100942 -0.079108365  2.388034624  2.365647290
#> [131] -1.236932233  0.185935753 -0.311668755 -0.951558672 -0.387693784
#> [136]  0.048727001 -0.353456439