Tukey's test for outliers

Perform Tukey's test for mild and extreme outliers.

outlier_tukey(
  x,
  mask = !is.na(x),
  threshold = c(1.5, 3),
  return.score = FALSE,
  ...
)

Arguments

x
A vector of data.

mask
A logical vector that defines which values in x will used when computing statistics. Useful when a subset of quality-assured data is available. Default mask is non-NA Values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

...
Arguments passed on to stats::quantile

Value

A vector the same length as x containing numeric scores or ordered factors.

Details

the values of threshold identify the multiplier of the interquartile range used to identify mild and extreme outliers. typical values are 1.5 for "mild" outliers and 3.0 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi
noise = rnorm(length(x), mean = 1, sd = 3)
y = sin(x) + noise
mask = noise < 1

outlier_tukey(y)

#>   [1] not outlier not outlier not outlier not outlier not outlier not outlier
#>   [7] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [13] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [19] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [25] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [31] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [37] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [43] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [49] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [55] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [61] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [67] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [73] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [79] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [85] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [91] not outlier not outlier not outlier not outlier not outlier not outlier
#>  [97] not outlier not outlier not outlier not outlier not outlier not outlier
#> [103] not outlier not outlier not outlier not outlier not outlier not outlier
#> [109] not outlier not outlier not outlier not outlier not outlier not outlier
#> [115] not outlier not outlier not outlier not outlier not outlier not outlier
#> [121] not outlier not outlier not outlier not outlier not outlier not outlier
#> [127] not outlier not outlier not outlier not outlier not outlier not outlier
#> [133] not outlier not outlier not outlier not outlier not outlier
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

outlier_tukey(y, mask)

#>   [1] not outlier     not outlier     extreme outlier not outlier    
#>   [5] mild outlier    not outlier     not outlier     not outlier    
#>   [9] not outlier     not outlier     mild outlier    not outlier    
#>  [13] not outlier     not outlier     not outlier     not outlier    
#>  [17] extreme outlier not outlier     mild outlier    not outlier    
#>  [21] mild outlier    not outlier     not outlier     not outlier    
#>  [25] not outlier     extreme outlier not outlier     not outlier    
#>  [29] not outlier     not outlier     mild outlier    mild outlier   
#>  [33] not outlier     not outlier     not outlier     mild outlier   
#>  [37] extreme outlier not outlier     not outlier     not outlier    
#>  [41] not outlier     extreme outlier not outlier     not outlier    
#>  [45] not outlier     not outlier     not outlier     not outlier    
#>  [49] not outlier     not outlier     mild outlier    mild outlier   
#>  [53] mild outlier    not outlier     not outlier     not outlier    
#>  [57] not outlier     not outlier     mild outlier    not outlier    
#>  [61] not outlier     not outlier     not outlier     not outlier    
#>  [65] mild outlier    not outlier     mild outlier    mild outlier   
#>  [69] mild outlier    not outlier     not outlier     not outlier    
#>  [73] not outlier     not outlier     not outlier     mild outlier   
#>  [77] mild outlier    not outlier     not outlier     mild outlier   
#>  [81] not outlier     not outlier     not outlier     not outlier    
#>  [85] not outlier     not outlier     mild outlier    mild outlier   
#>  [89] not outlier     mild outlier    mild outlier    extreme outlier
#>  [93] not outlier     not outlier     not outlier     not outlier    
#>  [97] not outlier     mild outlier    not outlier     not outlier    
#> [101] not outlier     not outlier     mild outlier    mild outlier   
#> [105] extreme outlier not outlier     not outlier     not outlier    
#> [109] not outlier     not outlier     not outlier     not outlier    
#> [113] not outlier     not outlier     not outlier     extreme outlier
#> [117] mild outlier    not outlier     mild outlier    mild outlier   
#> [121] not outlier     not outlier     not outlier     not outlier    
#> [125] mild outlier    not outlier     not outlier     not outlier    
#> [129] not outlier     not outlier     not outlier     mild outlier   
#> [133] not outlier     not outlier     not outlier     mild outlier   
#> [137] not outlier    
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

outlier_tukey(y, mask, threshold = c(2, 5))

#>   [1] not outlier  not outlier  mild outlier not outlier  not outlier 
#>   [6] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [11] mild outlier not outlier  not outlier  not outlier  not outlier 
#>  [16] not outlier  mild outlier not outlier  not outlier  not outlier 
#>  [21] mild outlier not outlier  not outlier  not outlier  not outlier 
#>  [26] mild outlier not outlier  not outlier  not outlier  not outlier 
#>  [31] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [36] mild outlier mild outlier not outlier  not outlier  not outlier 
#>  [41] not outlier  mild outlier not outlier  not outlier  not outlier 
#>  [46] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [51] not outlier  mild outlier not outlier  not outlier  not outlier 
#>  [56] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [61] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [66] not outlier  not outlier  mild outlier not outlier  not outlier 
#>  [71] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [76] not outlier  mild outlier not outlier  not outlier  not outlier 
#>  [81] not outlier  not outlier  not outlier  not outlier  not outlier 
#>  [86] not outlier  not outlier  mild outlier not outlier  mild outlier
#>  [91] not outlier  mild outlier not outlier  not outlier  not outlier 
#>  [96] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [101] not outlier  not outlier  not outlier  not outlier  mild outlier
#> [106] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [111] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [116] mild outlier not outlier  not outlier  not outlier  not outlier 
#> [121] not outlier  not outlier  not outlier  not outlier  mild outlier
#> [126] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [131] not outlier  not outlier  not outlier  not outlier  not outlier 
#> [136] not outlier  not outlier 
#> Levels: not outlier < not evaluated < mild outlier < extreme outlier

outlier_tukey(y, return.score = TRUE)

#>   [1]  0.137564109 -0.296800705  0.959831282 -0.341113079 -0.009074947
#>   [6]  0.281152317 -0.311796305  0.406824695  0.083736429 -0.132317980
#>  [11]  0.362512863  0.450556779 -0.084941229 -0.451938336  0.580620428
#>  [16]  0.308305176  0.991712060  0.046402278  0.071369288  0.126779405
#>  [21]  0.530680361 -0.102640880  0.082896719 -0.293202475  1.083224009
#>  [26]  0.949536962  0.084633527  0.244571988 -0.029839630 -0.217458879
#>  [31]  1.420212335  0.014088255 -0.247890053 -0.370824219 -0.209877746
#>  [36]  0.305581845  1.207017936 -0.257141444  0.717393995 -0.455949558
#>  [41]  0.402276671  1.043866148 -0.267828674 -0.206660839  0.208610871
#>  [46]  0.215200685  1.003555592  0.252692308 -0.285969361 -0.130211667
#>  [51]  0.012472702  0.794487462  0.067251510 -0.304764160 -0.275574036
#>  [56] -0.327004793  0.726282734 -0.159991927  0.232981196 -0.386067020
#>  [61] -0.114047330 -0.241040212  0.425691903 -0.254888860  0.131808399
#>  [66] -0.266542349  0.104699484  0.411940956  0.142789681  0.096706721
#>  [71] -0.108939654 -0.306479404  0.512840126  0.494520115 -0.043456308
#>  [76]  0.017713073  0.612732742 -0.419049980 -0.088439499  1.275308809
#>  [81]  0.191194992 -0.288126844  0.461613557 -0.354591920 -0.141140554
#>  [86] -0.281133785  0.207686321  0.447101285 -0.101232673  0.792825665
#>  [91]  0.035823228  0.887975394 -0.364062884 -0.263212662 -0.399444048
#>  [96]  0.343066574 -0.164322432  0.008251482 -0.083812728 -0.171875506
#> [101] -0.152407959 -0.424776120  0.047585830  0.000000000  0.866110151
#> [106] -0.402498411 -0.103534621  0.999528546 -0.287710840 -0.136598025
#> [111] -0.052421893  0.000000000 -0.184757611 -0.185585029  1.122165079
#> [116]  1.061850413  0.053943965 -0.323576350 -0.012637712  0.109589491
#> [121] -0.057303192 -0.049939117 -0.063804214  0.865581028  0.320636210
#> [126]  0.285205916 -0.112341700  0.539849828 -0.165642727 -0.019382725
#> [131] -0.370958772  0.051812293 -0.080095444 -0.238063231 -0.418884655
#> [136]  0.222555371 -0.134346079