Tukey's test for outliers

Perform Tukey's test for mild and extreme outliers.

outlier_tukey(x, mask = !is.na(x), threshold = c(1.5, 3), return.score = FALSE)

Arguments

x
A vector of data.

mask
A logical mask that identifies a subgroup of x to compute non-parametric statistics on. Useful when a subset of quality-assured data is available. Default action is to ignore NA values.

threshold
A length-two vector identifying thresholds for "mild" and "extreme" outliers.

return.score
if TRUE, return the numeric outlier score. If FALSE, return an ordered factor classifying the observations as one of "not outlier" (1), "mild outlier" (2), or "extreme outlier" (3).

Value

A vector the same length as x containing numeric scores or ordered factors.

Details

the values of threshold identify the multiplier of the interquartile range used to identify mild and extreme outliers. typical values are 1.5 for "mild" outliers and 3.0 for "extreme" outliers.

Examples

x = seq(0, 34, by = 0.25)*pi     
noise = rlnorm(length(x), meanlog = 1, sdlog = 3)
y=sin(x) + noise
mask = noise < 1

outlier_tukey(y)

#>   [1] mild outlier    not outlier     not outlier     not outlier    
#>   [5] not outlier     not outlier     not outlier     extreme outlier
#>   [9] not outlier     extreme outlier not outlier     not outlier    
#>  [13] not outlier     not outlier     not outlier     not outlier    
#>  [17] not outlier     not outlier     not outlier     not outlier    
#>  [21] not outlier     extreme outlier extreme outlier not outlier    
#>  [25] not outlier     not outlier     not outlier     not outlier    
#>  [29] not outlier     not outlier     not outlier     not outlier    
#>  [33] extreme outlier not outlier     not outlier     not outlier    
#>  [37] not outlier     not outlier     not outlier     not outlier    
#>  [41] not outlier     not outlier     not outlier     not outlier    
#>  [45] not outlier     not outlier     extreme outlier not outlier    
#>  [49] not outlier     not outlier     extreme outlier not outlier    
#>  [53] not outlier     not outlier     not outlier     extreme outlier
#>  [57] not outlier     not outlier     not outlier     not outlier    
#>  [61] not outlier     not outlier     not outlier     not outlier    
#>  [65] not outlier     not outlier     extreme outlier not outlier    
#>  [69] not outlier     not outlier     not outlier     extreme outlier
#>  [73] not outlier     not outlier     not outlier     not outlier    
#>  [77] not outlier     not outlier     not outlier     not outlier    
#>  [81] not outlier     extreme outlier not outlier     not outlier    
#>  [85] not outlier     not outlier     not outlier     not outlier    
#>  [89] not outlier     not outlier     not outlier     not outlier    
#>  [93] not outlier     not outlier     not outlier     not outlier    
#>  [97] not outlier     mild outlier    not outlier     not outlier    
#> [101] not outlier     not outlier     not outlier     not outlier    
#> [105] not outlier     not outlier     extreme outlier not outlier    
#> [109] not outlier     not outlier     not outlier     not outlier    
#> [113] not outlier     not outlier     not outlier     not outlier    
#> [117] extreme outlier extreme outlier not outlier     extreme outlier
#> [121] not outlier     extreme outlier not outlier     not outlier    
#> [125] not outlier     not outlier     not outlier     not outlier    
#> [129] not outlier     not outlier     not outlier     not outlier    
#> [133] mild outlier    not outlier     extreme outlier not outlier    
#> [137] not outlier    
#> Levels: not outlier < mild outlier < extreme outlier

outlier_tukey(y, mask)

#>   [1] extreme outlier mild outlier    not outlier     not outlier    
#>   [5] extreme outlier not outlier     mild outlier    extreme outlier
#>   [9] extreme outlier extreme outlier not outlier     extreme outlier
#>  [13] extreme outlier not outlier     not outlier     extreme outlier
#>  [17] not outlier     extreme outlier extreme outlier extreme outlier
#>  [21] mild outlier    extreme outlier extreme outlier not outlier    
#>  [25] not outlier     not outlier     not outlier     mild outlier   
#>  [29] extreme outlier not outlier     not outlier     mild outlier   
#>  [33] extreme outlier mild outlier    extreme outlier not outlier    
#>  [37] extreme outlier not outlier     not outlier     extreme outlier
#>  [41] extreme outlier not outlier     not outlier     extreme outlier
#>  [45] not outlier     not outlier     extreme outlier not outlier    
#>  [49] extreme outlier not outlier     extreme outlier extreme outlier
#>  [53] not outlier     extreme outlier not outlier     extreme outlier
#>  [57] not outlier     not outlier     extreme outlier not outlier    
#>  [61] not outlier     extreme outlier not outlier     not outlier    
#>  [65] not outlier     extreme outlier extreme outlier extreme outlier
#>  [69] not outlier     extreme outlier not outlier     extreme outlier
#>  [73] not outlier     extreme outlier not outlier     not outlier    
#>  [77] not outlier     not outlier     extreme outlier not outlier    
#>  [81] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [85] extreme outlier extreme outlier not outlier     not outlier    
#>  [89] extreme outlier not outlier     not outlier     mild outlier   
#>  [93] not outlier     extreme outlier extreme outlier mild outlier   
#>  [97] extreme outlier extreme outlier extreme outlier not outlier    
#> [101] extreme outlier not outlier     not outlier     not outlier    
#> [105] not outlier     extreme outlier extreme outlier extreme outlier
#> [109] not outlier     not outlier     not outlier     mild outlier   
#> [113] not outlier     not outlier     not outlier     mild outlier   
#> [117] extreme outlier extreme outlier not outlier     extreme outlier
#> [121] extreme outlier extreme outlier extreme outlier mild outlier   
#> [125] extreme outlier not outlier     extreme outlier extreme outlier
#> [129] not outlier     extreme outlier not outlier     extreme outlier
#> [133] extreme outlier extreme outlier extreme outlier not outlier    
#> [137] extreme outlier
#> Levels: not outlier < mild outlier < extreme outlier

outlier_tukey(y, mask, threshold = c(2, 5))

#>   [1] extreme outlier not outlier     not outlier     not outlier    
#>   [5] mild outlier    not outlier     mild outlier    extreme outlier
#>   [9] mild outlier    extreme outlier not outlier     extreme outlier
#>  [13] mild outlier    not outlier     not outlier     extreme outlier
#>  [17] not outlier     extreme outlier extreme outlier extreme outlier
#>  [21] mild outlier    extreme outlier extreme outlier not outlier    
#>  [25] not outlier     not outlier     not outlier     not outlier    
#>  [29] mild outlier    not outlier     not outlier     not outlier    
#>  [33] extreme outlier mild outlier    extreme outlier not outlier    
#>  [37] extreme outlier not outlier     not outlier     mild outlier   
#>  [41] extreme outlier not outlier     not outlier     mild outlier   
#>  [45] not outlier     not outlier     extreme outlier not outlier    
#>  [49] extreme outlier not outlier     extreme outlier extreme outlier
#>  [53] not outlier     extreme outlier not outlier     extreme outlier
#>  [57] not outlier     not outlier     extreme outlier not outlier    
#>  [61] not outlier     extreme outlier not outlier     not outlier    
#>  [65] not outlier     extreme outlier extreme outlier extreme outlier
#>  [69] not outlier     extreme outlier not outlier     extreme outlier
#>  [73] not outlier     mild outlier    not outlier     not outlier    
#>  [77] not outlier     not outlier     mild outlier    not outlier    
#>  [81] extreme outlier extreme outlier extreme outlier mild outlier   
#>  [85] extreme outlier extreme outlier not outlier     not outlier    
#>  [89] extreme outlier not outlier     not outlier     not outlier    
#>  [93] not outlier     extreme outlier extreme outlier mild outlier   
#>  [97] extreme outlier extreme outlier extreme outlier not outlier    
#> [101] extreme outlier not outlier     not outlier     not outlier    
#> [105] not outlier     extreme outlier extreme outlier mild outlier   
#> [109] not outlier     not outlier     not outlier     not outlier    
#> [113] not outlier     not outlier     not outlier     not outlier    
#> [117] extreme outlier extreme outlier not outlier     extreme outlier
#> [121] extreme outlier extreme outlier mild outlier    not outlier    
#> [125] extreme outlier not outlier     extreme outlier extreme outlier
#> [129] not outlier     mild outlier    not outlier     mild outlier   
#> [133] extreme outlier extreme outlier extreme outlier not outlier    
#> [137] mild outlier   
#> Levels: not outlier < mild outlier < extreme outlier

outlier_tukey(y, return.score = TRUE)

#>   [1]   1.749153089  -0.101161420  -0.030403162  -0.031613722  -0.254544501
#>   [6]  -0.064756419  -0.171683563   4.436087463  -0.225578545  10.891861872
#>  [11]  -0.083574479   0.296126779  -0.249474253   0.042970027   0.059041726
#>  [16]  -0.392152795  -0.004718403   0.437654274  -0.306815581  -0.341549513
#>  [21]  -0.176347341 331.270422082 151.131271092   0.048776596  -0.094148977
#>  [26]  -0.053352838  -0.031302171  -0.103664234  -0.279673746   0.010261923
#>  [31]   0.051141124  -0.119923461  22.587050674  -0.149188978   0.158454742
#>  [36]  -0.019841357  -0.142818051   0.042687808   0.048531942  -0.286487075
#>  [41]   0.945023899  -0.009638931  -0.046450547  -0.253269623   0.016188029
#>  [46]   0.039204502  72.215321603   0.042072949  -0.435428442  -0.013605183
#>  [51]   9.778396329   0.572193848  -0.014807206  -0.341911740   0.061330111
#>  [56]  29.264958587   0.015892422  -0.011528541   0.064227251  -0.100594620
#>  [61]   0.020521695   1.492513286   0.009851493   0.007057487   0.004426605
#>  [66]   1.087094890 178.531479255  -0.169527344   0.018127823  -0.309926126
#>  [71]   0.058079476  43.838717679   0.000000000  -0.254872985  -0.027950291
#>  [76]  -0.023503360   0.019888565   0.036507641  -0.298048844   0.033945216
#>  [81]  -0.437359275  14.909371999   0.584007511  -0.141052184   0.000000000
#>  [86]  -0.434771093   0.060636345   0.031884925   0.125934759  -0.053532913
#>  [91]  -0.049828374  -0.110116688   0.012141036  -0.219590405   0.984217386
#>  [96]  -0.142666783  -0.348356216   2.244532878   1.161625493  -0.033228452
#> [101]   1.021270072  -0.084196602   0.059353142   0.048400898   0.012537284
#> [106]  -0.364339933  14.156980011  -0.221928244  -0.051695113   0.049375795
#> [111]   0.053219870  -0.126152955   0.019664940  -0.047973148  -0.053362348
#> [116]  -0.128613942   6.605086672  14.359559275   0.033460673  14.741979347
#> [121]  -0.484100333  22.471580564  -0.280029831  -0.119963646  -0.411108971
#> [126]   0.040599338  -0.464976234  -0.446370840   0.011052352  -0.293287795
#> [131]  -0.054953462  -0.227081290   2.896136277   1.347760973  42.385986899
#> [136]   0.001069021  -0.261973619